[roll] Roll fuchsia [netstack3] Remove DeviceId from network sim
The `DummyNetwork` simulation in the `context::testutil` module has
`SendMeta` and `RecvMeta` parameters for the metadata produced when
sending and consumed when receiving, respectively. It also has a
`DeviceId` parameter. This latter parameter is redundant - when a
network is being simulated at the level of an entire stack complete with
devices, `SendMeta` represends the device ID of the outbound device on
the sending host, and `RecvMeta` represents the device ID of the inbound
device on the receiving host. (In other scenarios, such as simulating
communication between two transport-layer sockets, there is no notion of
a device). `DeviceId` is redundant, and so is removed in this commit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648029
Original-Revision: b1ebb7c64a9d7524c7fccd9e1239757e41c6dd1b
GitOrigin-RevId: 08e34f04c2c92280acee6ac6ea61aa7dd1c82852
Change-Id: I5484b921dbee904eb85b7befd0e57552d58adf8c