[roll] Roll fuchsia [wlan] Remove use of base frequency in APIs
This change removes the base frequency in each band capability. While
it's technically possibly for base frequency to differ between
implementations of the same PHY mode, in practice, channel numbers and
channel frequency indices in the context of particular industry
standard bands, such as 2.4 GHz and 5 Ghz, have the same one-to-one
relationship in every regulatory domain and in the absence of one.

Test: Manually verified Atlas and Astro connect and transmit data.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640281
Original-Revision: 3d5718154e38d11f7b389ae4bddf524da16efd03
GitOrigin-RevId: 5b2c7b97b848d5872a4431fb527c6745be2b0d15
Change-Id: I34625b79c276e64f7910e3a91dc6de48ff5891c4