[roll] Roll fuchsia [netstack3] Do not set IPv6 flow label
As per RFC 6437 section 2,

  The 20-bit Flow Label field in the IPv6 header [RFC2460] is used by a
  node to label packets of a flow.  A Flow Label of zero is used to
  indicate packets that have not been labeled.  Packet classifiers can
  use the triplet of Flow Label, Source Address, and Destination
  Address fields to identify the flow to which a particular packet
  belongs.  Packets are processed in a flow-specific manner by nodes
  that are able to do so in a stateless manner or that have been set up
  with flow-specific state.

Netstack3 does not use the flow label or keep any state for a flow so
this change removes this for now to make future refactors easier.

As we can see in this change, little work is needed to generate the flow
label so adding back this feature when there is a need won't be a
significant burden.

Original-Bug: 94089

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647339
Original-Revision: 8d5577457f8e8cbcbf1b3c4cc021f22eda193011
GitOrigin-RevId: 088712c8d81aa5262d6b518b3bf8328d586bf2f2
Change-Id: I4ac53b25d4a05845269cb0c89dd464ef99963d87