[roll] Roll fuchsia [fidl][llcpp] Remove header from Wire_<T> types
The template types `fidl::Wire{Request|Response|Event}<T>` have a
private, 16-byte _hdr field, for storing the header message portion of
a transactional message (ie, a concatenated pair of 2 FIDL messages). To
enable top-level union and table payloads in C++, we'd like to be able
to generate something like `WireRequest<T> final : public T` for cases
where T is a union/table payload (and perhaps, in the future, for the
existing struct case as well). Because users should not be using these
types to access header information anyway, it is desirable to make the
layout of `Wire_<T>` to match that of `T`. Instead, FIDL-internal
methods that need both the header and the body will use the newly
created `fidl::internal::Transactional_<T>` template, which will wrap
the now headerless `Wire_<T>`.

This change accomplishes the stated goal in a manner that mostly does
not touch public uses of `Wire_<T>`. There are two categories of
exceptions:

* There were some cases where end users (especially tests) were using
  the non-public `_hdr` field. These have been removed.
* There are a small handful of cases where methods like `Unwrap()` were
  called on body-less messages. Since body-less `Wire_<T>`
  specializations no longer emit such methods (and, in the `Callback`
  case, do not provide a pointer to the body as a parameter), this small
  handful of clients needed to be updated.

Test: fx test
Original-Bug: 88343
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648011
Original-Revision: 4e617f41295fb82666d2691846de256c92453e89
GitOrigin-RevId: 8fae2118a86ebe16f5836c75a4e81278e754fe6e
Change-Id: I863c25e019a3bd22c73f89fdf7c91732603fe5d7