[roll] Roll fuchsia [vfs] Support handling vmo dirty requests in PagedVfs
This CL allows PagedVfs to control the transition of the VMO page state
from clean to dirty. It introduce a method, PagedVfs::VmoDirty(), in
which filesystem can receive requests for the transition.
If the transition is acceptable, filesystem can handle it by doing some
specific works.
In response to the request, it should request ZX_PAGER_OP_DIRTY so that
kernel transit the pages to dirty and wake up waiters.

Test: fx test fs-connection-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645261
Original-Revision: b757a9f2386b421b2cd56fcc06f4bb156ac3ca21
GitOrigin-RevId: 95abff09e1ef49df2752a50b01fc3ddc6975920b
Change-Id: Ie39a05dc514da9805f68cb4fda52a1370ea87b2e