[roll] Roll fuchsia [f2fs] Add a writeback policy
This CL introduces a Writer class that merges Pages flushed from
PageCache, and a caller can schedule a task to Writer to flush
the merges Pages to disk in a synchronously or asynchronously manner.

It also defines a new writeback policy. If the number of
committed pages exceeds a soft limit, f2fs flushes dirty Pages
as many as a segment. When the enough number of Pages are merged
in Writer, f2fs schedules a Writer task to write out the Pages on disk.
Meanwhile, when the amount exceeds a hard limit, it writes out all
dirty Pages for all dirty vnodes synchronously. To measure the number
of committed pages, f2fs tracks the number of dirty/writeback Pages
that keep committed pages with vmo_op_lock. The limits are adjustable.
When it traverses FileCache to find dirty Pages, it also releases clean
Pages. It helps to avoid the high handles warning.

Test: fx test f2fs-slow-fs-tests f2fs-fs-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626194
Original-Revision: 6af340004699a560f240506ba65fd15e4f6edd96
GitOrigin-RevId: b1749169fb796a2a48536053d725b317af7308f5
Change-Id: I7e8c6a0393c4c993acfeb94bb33a8c2be9f909a7