[roll] Roll fuchsia [f2fs] Do not acquire a tree lock when releasing Pages
This CL keeps Pages as weak pointers in FileCache, which allows to free
the resource of a Page in fbl_recycle() where no one refers to the Page.
It eliminates the need for acquiring the tree lock to track references
to Pages every release. It also reduces the locking cost at Page
invalidation.
With this CL, we can see that the slow-fs-tests time is reduced by ~20%.

Test: fx test f2fs-slow-fs-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641967
Original-Revision: 9a86a7246306c5365f8a67cbff94d5b558c0cb66
GitOrigin-RevId: 19ac11636e4e779e01a41e3fcaeee1c93002d7c0
Change-Id: I8ac2d5252f6948a0f7eb0f5d61f5ed1969b1d097