[roll] Roll fuchsia [x86][acpi] Move GetBti to FIDL.
We switch from allowing devices to specify an arbitrary PCI
bus/device/function to get an IOMMU for to determining one for
ourselves.

The new GetBti implementation adds some constraints over the old one:

* Callers cannot be on a bus (i.e. they must only be ACPI devices). We
  will relax this restriction in the future, most likely to only prevent
  PCI devices from getting an ACPI BTI.
* Callers will always get back a dummy BTI with a meaningless (but
  unique) BTI ID.

In the future we will add support for determining the IOMMU to be used
for an ACPI device based on the DMAR entries.

Original-Bug: 80244
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639014
Original-Revision: 66287be0111290e77ea186d8bb8990f58c8fcf83
GitOrigin-RevId: c38b778ce1cdf55248db42b112672d6070a6df15
Change-Id: Ibd4755cabd6c56de0f2e9065ff72905810bd951f