[roll] Roll infra/recipes [clang_toolchain] Disable the use of zlib and libxml2 for stage 1
We only build zlib and libxml2 once for the target platform, but when
cross-compiling Clang to another target, stage 1 and 2 are built for
different platforms (stage 1 for host, stage 2 for target) and so we
would get an error when trying to use zlib and libxml2 for stage 1.

We could build zlib and libxml2 twice, once for host and once for target
platform, but it's not clear whether we even need these for the stage 1
since it's just a throwaway compiler used to build stage2 and we don't
use compression in stage 2 build at the moment, so disable their use
for now.

Original-Bug: 94203
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/649021
Original-Revision: 9ed253bded40b8d9d30b1aa623a30d9233642b66
GitOrigin-RevId: 8c4be3ea428486250bb921ef267201144406ce15
Change-Id: I992a548709f1c5b70bbbd5cd5002d7a9a9b4455a