[roll] Roll fuchsia [fdf][compat] Store Driver in Device
The zx_driver_t* pointer that a DFv1 driver hands to us
is global across the drivers of that type in a driver host.

We cannot use it to add a device, since it will add the
device to the most recently created driver of that type,
instead of adding it to the correct driver.

This is tested by having two instances of a driver each
add a device with the same name in the same driver host.

This also adds a test to make sure globals work correctly
within a DFv1 driver.

Original-Bug: 94230

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648041
Original-Revision: 645ec7d368bae360d84d6ed3d5b8b1a0c09db60a
GitOrigin-RevId: 3e279c912cc5ef51cb131209ee267faabf3cf59a
Change-Id: I6bebbb45e262a668623ecfcff81f76fba528c8d3