[roll] Roll fuchsia [debugger] Fix a bug in DebuggedProcess::OnResume
The IPC specifies that in ResumeRequest, an entry with a process koid
and a 0 thread koid will resume all threads of the given process.
However, this is not correctly implemented in DebuggedProcess::OnResume
and causes a problem when debug_agent detaches from a process.

Original-Fixed: 90060
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645116
Original-Revision: 33a4163de44172bd8d1f724250d90211fe21637c
GitOrigin-RevId: 7f4d53f32420415352bacae4830ee62c2131b5b6
Change-Id: If2a299872ba397bbff713c63cf5be0fdd8d484a4