[roll] Roll fuchsia [pm] pm_publish_archive
This change implements a new gn template: `pm_publish_archive`.
The template follows the same procedure as `pm_prepare_publish`
combined with `pm_publish`, but produces a zip file as output,
rather than a repository directory of files. This template can be
used in combination with the `zip_resource` gn template to include
a fuchsia packages TUF repository in a fuchsia package. This is
useful, for example, to implement hermetic package delivery tests
that interact with an on-device package server.

Original-Bug: 88453

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640725
Original-Revision: 2d2474d14c55844ad765246e504cf70fbee896c3
GitOrigin-RevId: a79e431f5b1fb9487d5c8fa809fc81903f3cb738
Change-Id: I05d6df0f6494dce3748db47a76b5fd2bdadf4d6d