[roll] Roll fuchsia [fidl] Template NaturalMessageEncoder on Transport
In an upcoming CL, NaturalTransport will need to store handle metadata
that is specific to a given transport. In order to facilitate this, this
CL templates NaturalMessageEncoder on Transport.

NaturalBodyEncoder is not templated because it should be ok for the
forseeable future to use the ChannelTransport with it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647925
Original-Revision: 47769958a6111c016376d66d949fb47b93c7a18a
GitOrigin-RevId: 5212f966b38a5d398b4d050b03eaf3f60a50947f
Change-Id: I91a228eb55d7d0cf66910ee6b3fefc96ae8d935d