[roll] Roll fuchsia [zircon][ldso] Avoid infinite recursion in dlsym
It's possible to infinitely recurse in find_sym_for_dlsym if DSOs
have circular dependencies. This dynamic linker code will be
eventually replaced. A hack we can do now to allow clang to roll is
dynamically check which DSOs we've visited already and if we hit a
cycle then we immediately fail the lookup.

Original-Bug: 93847
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647643
Original-Revision: 323922a50fc348618a1334c49324585ba5f2822e
GitOrigin-RevId: 287e5440e36942060235606122e0523291691ef3
Change-Id: I9bbd17858d0d68a98f7e4597ba00eb5674cd2458