[roll] Roll fuchsia [build] Make hermetic action failures less verbose
If a build step that's wrapped by the action tracer fails in a generic
way (not related to action tracing) the error message would previously
include a stack trace from the action tracer script in addition to the
forwarded stdio of the failed command, e.g.:

  <full action tracer command line>
  <stdout/stderr of failed command>
  Traceback (most recent call last):
    File "../../build/tracer/hermetic_inputs_action.py", line 59, in <module>
      sys.exit(main())
    File "../../build/tracer/hermetic_inputs_action.py", line 48, in main
      subprocess.check_call(cmd_args)
    File "../../prebuilt/third_party/python3/linux-x64/lib/python3.8/subprocess.py", line 364, in check_call
      raise CalledProcessError(retcode, cmd)
  subprocess.CalledProcessError: Command '[<cmd>]' returned non-zero exit status 1.

(where <cmd> is the command line that's passed to the action tracer, and
is redundant because it's already included in the full action tracer
command line).

This just adds meaningless noise to the error message and makes it
harder to determine the actual cause of the failure, so we will now no
longer display that stack trace.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649623
Original-Revision: 10f8c6a26d75ce11f3e81edb9fa5356d18f56bc2
GitOrigin-RevId: bbf32c47d854575a993484f0ffdb817d4ea23f6c
Change-Id: Ica6e0d1f0548e3c0cac021ef91d34bb230a58843