[roll] Roll fuchsia [wlan][softmac] Use operating_channel_list field name
The use of the operating_channel_list field replaces
fuchsia.hardware.wlan.phyinfo.WlanInfoChannelList. The new field name
better represents the data it contains which is a list of operating
channels in the sense that IEEE Std 802.11-2016 defines "operating
channel": The operating channel is the channel in which beacons are
transmitted.

Additionally, this change removes ValidateWlanSoftmacInfo which only
checks that each channel number is in ranges defined in
Annex E (even though he channel numbers defined in Annex E apply on
when in the given regulatory domain). However, it's really up to the
driver whether a channel number is valid operating channel for a band
or not.

Original-Fixed: 91033
Test: Manually verified Astro and Atlas still connect and transmits data.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642484
Original-Revision: 6a85d0c092071c4f1d737adc2d4d9feced952fd6
GitOrigin-RevId: 32f56945b71a37520ffe549c968d79c6b8f0df16
Change-Id: I5a1ce00d99e3a814f77f467b31aea28415a5fbb8