[roll] Roll fuchsia [fshost] Try to give at least 10% to /data
When we format a custom filesystem, we arbitrarily give it 24MiB
(100MiB for f2fs), which is a fairly small allocation, especially if the
disk is substantially bigger.  For some filesystems (fxfs, f2fs), this
is the final partition size, as they cannot grow.

Tweak the algorithm to give the greater of 10% of the disk or 24MiB.

Original-Bug: 93060
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649192
Original-Revision: c95e72ffa645548dffef3ea1a1ef974081816c85
GitOrigin-RevId: ddb92325092f422010230965e6a7ddec0d6f9276
Change-Id: I27a683d5567f0118a8e5177f31038e6b91d20d10