[roll] Roll fuchsia [bt][testing] Use const duration for integration test timeout
`zx::Duration::from_second` is a const fn, so it can be used to create
a zx::Duration const. Do this to define the integration testing timeout
instead of defining a separate method to return a new Duration each
time the timeout is referenced.

Original-Bug: 94137
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648381
Original-Revision: 88087aef12070ad8b8a006558d703dc2504b9464
GitOrigin-RevId: 27f9f64863dbf3a7793ea6e137356efd8cee7658
Change-Id: I71076c40b5a0f52a77b0d8f64b04dd671aee9789