[roll] Roll fuchsia [ffx] Add UpdatedTarget event to improve target waiting.
This adds an UpdatedTarget event to improve target waiting behavior,
specifically if awaiting a manually added target for a specific nodename
that is known a priori.

If we manually add a target and then run the wait_for_target function,
we will receive NewTarget events for each new target seen (including
synthesized events). However, if we then make an RCS connection and
identify the target, we will not succeed in waiting for this target as
there is no indication that it has been updated.

Original-Bug: 94317
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649941
Original-Revision: 8d7942e4c75945fd9048156cce0a68ef8e6a25bb
GitOrigin-RevId: d141205d5c362698b35140c8afe9c5c8b648d172
Change-Id: Ibe22bb8a22023b752d8938a38dcfc96fda9a21ec