[roll] Roll fuchsia [ffx][config] Fix tests overwriting user settings
ffx config was designed such that if it was run in a test, it would only
write to temporary data. While this was true for the environment config,
ffx_config::check_config_files would check if the environment contained
a path to a user config. If not, it would set it to the real default
user config. This would result in tests writing to the real user config.

To avoid this, this patch changes initialization of the test config to
create a temporary user config file, and point the environment at it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650061
Original-Revision: b9080619af854420f9a5089425ed2988b445a19f
GitOrigin-RevId: 0047228949cf25ba22f1ed0250423e6674863b08
Change-Id: I394da5520fb7d996b1b66289090d21b28fa3f543