[roll] Roll fuchsia [scenic] Removes the command line flag that forces flatland use
The flag was introduced during the transition from the gfx
APIs to the Flatland APIs use for graphics, to ensure that
the tests always execute using Flatland.

Now that it seems the Flatland API is the default and we
are not going back, let's remove this unneeded bit of
complexity.

Original-Fixed: 92930
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650503
Original-Revision: 5af0ec49d87e35a3922ea45714fd66a9fa1ce4dd
GitOrigin-RevId: da6560986c4913f2b5ab54762c549f4e186c21b9
Change-Id: Ib3ac4a2b0d02c8367e41135d16792378bca7d47d