[roll] Roll fuchsia [audio][virtual_audio] virtual_audio cleanup
The virtual_audio driver receives a bit of code cleanup --
largely conversion of DEBUG_ASSERTs and return-error to ASSERTs.
The thinking is (1) for aspects that we CAN control, we do not
intentionally configure things to cause errors, and (2) for
aspects that we cannot control, we want to surface problems
as soon as possible.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579460
Original-Revision: fd6683ef68a3fb9b4e805578c833334b2842e0f5
GitOrigin-RevId: 5f7d3ca5dbe58945b19e0e091b6ce8c4912f9670
Change-Id: I21f68b26794c476a3af7584fdd6debb5f8a12fe7