[roll] Roll fuchsia [input] Adds a display ownership handler to the input pipeline
This commit allows us to stop processing input when the user jumps away
from the product, as in the case of jumping to and back from
the virtcon.

While this commit will multiplex virtcon and product, I'm still not 100%
pleased with the way it is structured. But to resolve that part
completely, I'll need to do some additional work that this commit does
not need to wait on.  Hence, added some TODOs in the change and sending
this one for review.

Original-Bug: 70249
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648201
Original-Revision: 04631e92e9f5deda80b7655988ae4c622aeb16ab
GitOrigin-RevId: e74d8e460ae2edfb4a5accbf258f2a1243581152
Change-Id: I6ce6334964363b91358b84ad91d3b29860220fee