[roll] Roll fuchsia [mdns] establish pattern for nonce service implementations
This CL moves the implementation of the ServiceInstanceResolver
service out of MdnsServiceImpl and into its own class, each instance
of which serves a single client.

The FIDL rubric requires that, when a client makes an avoidable error
(such as passing an invalid argument), the service implementation
should log a message and close the channel. This prevents us from
using any service implementation that serves more than one client
(because we have no way of knowing which connection to close).

Here we establish a pattern for nonce service implementations. The
ServiceImplManager class manages lifetime for service implementation
instances. The manager employs a deleter callback, passed to the
instance's constructor, which the instance can use to request
deletion. ServiceImplBase is a base class for the service
implementations handling binding, deletion, and connection errors.
It also supplies the instance with access to the Mdns class, which
is an in-proc API that the instances call do get work done.

This CL also changes ServiceInstanceResolver to handle client errors
as described. There are currently no clients out-of-tree, so this can
be done without breaking.

Test: no behavior change, unit test added

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650807
Original-Revision: 414d08762c0d34f2e2a4bdbbca4b2268c9b122bc
GitOrigin-RevId: b7222afc4497cc8f47b2ce2453f5a4f7db3a9691
Change-Id: I0ea73066463f34f7dd10d1cffbd34e44858e0dab