[roll] Roll experiences [e2e] Disable flaky e2e test on Atlas
This has been flaky for months and frequently closes the tree. Disable
it (but leave it running on FYI builders) until an owner has time to
debug it.

It's mostly flaky on Atlas, so it can remaing blocking on other device
types.

Original-Bug: 79921
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/650923
Original-Revision: b9122ffbae205e7a99dc10e0158b282352fe573c
GitOrigin-RevId: fdee3c0cc944ea4d9222fe50b97da770d826cfce
Change-Id: I225739daff4990a6da55b9db15a08180da5d8a1b