[roll] Roll fuchsia [check-licenses] Results package
This handles saving NOTICE files, printing a summary to the command
line, and producing a manifest / summary of all the license texts that
were found in the build. Basically, all outputs for the tool should come
through this package.

We can move this back up to the base check-licenses package in the
future if it makes more sense.

Also added a driver file to check-licenses to make it easier for me to
test the v2 codepath. This code is still not being run by default.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650826
Original-Revision: 9a9d5ec656e0d274712e7da567b82c7ee202fa4f
GitOrigin-RevId: 74e5c6815d2a79067aa5cbdb8c9c4f929152ce07
Change-Id: If2a5d4f409587b78606c5c8e36da8c4c97ab16b5