[roll] Roll fuchsia [svc] Add sub directory support
This CL adds two new functions to the `svc` library,
|svc_dir_add_directory| and |svc_dir_remove_directory|,
that will allow the hosting of fuchsia.io.Directory-backed
directories in the |svc_dir_t| root.

This is needed to enable support for equivalent functionality
in a downstream client of this library.

Original-Bug: 89036

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649781
Original-Revision: ff93faef24d7080921510baaaaaeb779cfdee3c4
GitOrigin-RevId: 2ff0890ef8bcf335e0a6e295b1431bce8bbfd358
Change-Id: Ia836a68665f1e1e6ed2459064f6867b85faea75a