[roll] Roll infra/recipes [fint] Locate clang crash report reproducers
This reimplements logic that's currently in recipes to locate clang
crash report reproducer scripts, before uploading them to GCS:
https://fuchsia.googlesource.com/infra/recipes/+/739536bcf92e7095658cafab5b90aa759b052c6c/recipe_modules/build/api.py#1174

I had wanted to move that logic out of recipes for a while because it
encodes more knowledge of the output directory structure than is
appropriate for recipes, but I'm doing it now to set a precedent for
having fint emit generic debugging files whenever a build fails that the
recipe will upload to GCS (corresponding recipe change upcoming). This
will be useful for uploading various ninja log files to GCS to help
debug tricky incremental build failures.

The clang crash report recipe logic tars and gzips the files before
uploading them to GCS, but I didn't bother to do any compression here
because:
- Emitting a tar.gz would add significant complexity to the code
- The files are normally relatively small (single-digit megabytes)
- We only upload them on a small subset of failed builds
- There will generally be a TTL on the files in GCS

So all in all, I didn't deem it worth it to implement gzipping. We can
add that later if bandwidth or storage does becomes a concern.

Original-Bug: 94334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650355
GitOrigin-RevId: cc2011d0aca3e74df1bdcba14acdccaa26457cec
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/651306
Original-Revision: 62dcaefa099e8604ccb9367f2293892936f41519
Change-Id: I7896f503184976b0a2ebbb1e07afed05d6f06e05