[roll] Roll fuchsia [fdf] Fix UB while removing node.
Node::Remove() would call Remove() on all of its children, which would
mutate the children list of the parent. This would result in undefined
behaviour because the vector was being iterated on when it was mutated.

This CL switches the children_ member to be a std::list, and slightly
modifies the Remove() loop so that we don't trigger UB.

Original-Bug: 94561
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651286
Original-Revision: 6f8f15b44b46aceafb5f05515a31179d8d8bda74
GitOrigin-RevId: f3536abeac22209bd5cc12264b7c282db842d06c
Change-Id: I3547ffaf15ddff737b9d6ae3dd9a8f51a3d2d70d