[roll] Roll fuchsia [element_manager][annotations] Introduce WatchAnnotations
This CL introduces the `WatchAnnotations` FIDL protocol method, and
implements the function for `felement::AnnotationController` and
`felement::Controller`.

The `watch_annotations` function utilizes the `async_utils::hanging_get`
API to perform a hanging-get so that clients can watch for annotation
updates.

A `WatchSubscriber` is initialized for each client stream to keep track
of the client state in regards to annotation updates. When
`watch_annotations` is called for the first time by the client, or if
the annotations have changed since the client's last request, the
current set of annotations are returned immediately. Otherwise, the
request will "hang" and a response will not be returned until the
annotations have been updated.

The reason for this change is to provide a mechanism for a client to
synchronize its annotations set with updates. This provides feature
parity with `StoryPuppetMaster.WatchAnnotations` in Modular for
`GraphicalPresenter`.

Original-Bug: 56594
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647783
Original-Revision: d3acdbc95aa632c9b142be6f306303c562b4f4e7
GitOrigin-RevId: 9a5f3065fbec08dc28ec0a86d3a28eb75060817e
Change-Id: I1b398e59ac7514eb1148153c4222871097890029