[roll] Roll fuchsia [input] Reworks the display ownership handling
I wasn't all too happy with the previous approach to handling events
because it didn't handle display ownership loss and gain at the time the
change happens, but rather at the time the next input events comes in.o

I reworked the approach, centralizing all event processing in a single
loop, and using async channels to funnel events in and out of the
handler.

What remains is a rename of the struct and files to not use the word
"handler", since in the input pipeline it has a very specific meaning.
I opted to rename in a separate change to simplify the resulting diff.

Original-Bug: 70249
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650103
Original-Revision: f12761c2e04ec376ce6252a9ec6729c3ee13f012
GitOrigin-RevId: 650e971d97eb05a4c1f065fcc0b0080c4c8b1287
Change-Id: I26ff78c5ef74e162b06b50520f17d16af24370d1