[roll] Roll fuchsia [blobfs] Remove kDefaultJournalBlocks, use kMinimumJournalBlocks.
Both constant were the same value because the minimum value was quite
large to accommodate the largest operation. This removes all the logic
that tries to intelligently shrink/grow the journal size based on space
available, since it was previously making a larger journal for smaller
partition sizes. This unexpected behaviour was creating a mismatch in
the expected size of the filesystem in the host tool mkfs.

Had to update one test's values since it seemed to be just barely
passing due to this miscalculation.

Original-Fixed: 94138
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651008
Original-Revision: 565b28213c6c70696118bbf346c10bc395affb21
GitOrigin-RevId: 53f7ae33294fc6fbccf9c5dea52a5fc96ded3018
Change-Id: If8b8510b2ed5479bf287d93a07c708d82ab8e8c0