[roll] Roll experiences [workstation][chromium] Deflake workstation_chrome_advanced_smoke_test
The test started to fail at opening the app launcher.
https://ci.chromium.org/ui/p/turquoise/builders/global.ci/sucrose_eng.chromebook-x64-debug-e2e/b8820983292116563057/overview
This isn't observed in local test, but is on the builders. I added the logic to repeat opening the app launcher and look for the Chromium app entry for a fix time duration as the flaky might be coming from the delay in completing a task on builders.

Original-Bug: 94042

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/651581
Original-Revision: 8fe5c44cbcebec0c3dbd868e1577c31252e198c8
GitOrigin-RevId: d844caf899ae898e43db21a2f2854580ed5b2b9c
Change-Id: I4c982cd51fb90293cff79ec19b032b072a6fdd24