[roll] Roll fuchsia [vm][test] Handle VMO resize
Remove ZX_VMO_RESIZABLE from the default creation option in pager core
tests.

Modify the VMO resize helper to also change the mapping when the VMO is
resized. This is needed to be able to access the VMO safely after a
resize.

No change in behavior with this CL. These changes are needed for tests
that will be added in following CLs.

Test: core-pager
Original-Bug: 94371
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650841
Original-Revision: 85d3ecc1e96ace727e6759e04492105119eb1035
GitOrigin-RevId: f247e3ec6a07c9acbfc0642efed3d3d39f467f18
Change-Id: I901300f2addb2d2c643339891602ddf8efb7811a