[roll] Roll fuchsia [lowpan][openthread_fuchsia] Fix for Platform UDP Sockets.
There were two problems with the existing Platform UDP sockets:
The first is that the socket was being opened as an IPv4 socket
instead of an IPv6 socket. The second was that in error conditions
when sending we were double-freeing the message. This CL addresses
both problems.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651722
Original-Revision: 309e893ef8d8648465fd5b844afb13e1cb3c42a0
GitOrigin-RevId: 69ef0c37cb28104b4ef1a0bcf9986aba92818ebe
Change-Id: I6a33622b264483d7c7be55c4084c080bc9cb9282