[roll] Roll fuchsia [Recovery] Fix button race conditions.
Button presses were not recorded unless matched in a match
statement. They are now recorded and propagated as a plain
button press.

There was a race condition when in waiting mode when a cancelled
event could get through because of a background task in main.
This is now ignored.

Rustfmt did a lot of reformatting.

Original-Bug: b/215578671
Fixes: b/215578671
Test: New test added, fails before fix
Test: One test removed, it is no longer relevant
Test: All tests run (the same three fail that failed before)
Test: Verified that it no longer panics when both buttons released.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634723
Original-Revision: 51fb09ce994627d016855243dc656c2c304ae700
GitOrigin-RevId: af61d9c4c8eacd21fdc262a3721cd706b271d8d7
Change-Id: Ib0c794b537b6552242b02761d7676ba341558ac7