[roll] Roll fuchsia [size-checker] Write the report to file.
By default the size_checker tool writes the report to stdout. When `--report-out path` is specified the report is written to a file, and only errors are written to standard stream.

This makes possible to run the tool as part of the build system, where nothing should be printed when the tool runs without error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642801
Original-Revision: cba16666edbc3d0e04608aba0c48fefee3721c8d
GitOrigin-RevId: b3d1cf17122405dfa86b2679c4d3dac409d7f8e1
Change-Id: Ia90b89b738d7cf62eb94c9a47e6609346fd3181a