[roll] Roll fuchsia [svc] Split `svc_dir_create`
This CL introduces a new way to construct `svc_dir_t`
that splits up the creation of the object with it serving
on a channel. This is done in order to align with expectations
of downstream clients that enforce that the outgoing directory
is populated *before* it being served.

The goal is change `svc_dir_create` to solely create the object.
However, since that changing the function's semantics may break
its clients. The API changes will be done through a soft transition.
To that end, this CL introduces two new functions,
`svc_dir_create_without_serve` and `svc_dir_serve`. The former
creates the object, while the latter serves its content on the
provided handle.

`svc_dir_create_without_serve` is only a temporary name that will
be used only for the duration of the migration of `svc_dir_create`
clients to it.

Original-Bug: 89036
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651786
Original-Revision: 44a5fb51368eca6349b0175b5870539cfe740a8f
GitOrigin-RevId: 00c1ca9a627004eed3be97ac5a7bb5fac2e219fe
Change-Id: I69dcb60a79b428eb7bf2c2b67ec64c219630aeed