[roll] Roll fuchsia [bsdsocket_test] Stringify cmsg type
Instead of having to manually translate cmsg level/type to a string, use
a macro to do it for us. It has the benefit that we do not need to add a
new switch case every time a new cmsg is added.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651421
Original-Revision: 83de69fac3f47d9347edc4f6b9d30bd10723fe53
GitOrigin-RevId: 6b09328b4d5989513526dcdf13af478da76d07ad
Change-Id: Id7dcd69265e370db920e9eb8084e24d7440a62f9