[roll] Roll fuchsia [kernel] Re-name and change is_user_address[_range]
This updates the behavior of is_user_address to instead reflect
that the user can access this address rather than the address is
within the range of USER_ASPACE_BASE and USER_ASPACE_BASE +
USER_ASPACE_SIZE. On aarch64, this is a check that bit 55 is zero.
On x86_64, this is a check that bits 48-63 are zero. The function
is renamed to `is_user_accessible` to indicate that this address
is accessed in userspace.

Note that before on x86_64, the user aspace would end one page
earlier to avoid being able to map things in the last page. Code
that relies on that behavior should still work.

This also updates some tests with the new implementation. Some
previously "invalid" addresses used in tests are updated to reflect
"invalid" addresses with the new logic. In particular, some test
addresses used in the user_copy tests may allow the copy to actually
happen. For those tests, we just accept the page fault but the result
status should still be the same.

Original-Fixed: 92770
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639991
Original-Revision: 9a505119042d8bc0077338d45640777457449bd7
GitOrigin-RevId: 59a54862c3bb9b42c907ba9ea7c86a04b4e63516
Change-Id: I2a9098d93c63bd98106e5823ad64ce1ef605e602