[roll] Roll fuchsia [system-update-checker] Fix flaky tests
By virtue of the test fakes running quickly, the tests that verify that
the status was committed succeeded. By adding a timeout, these tests
failed because they check that an update was started when they should
check that an update has completed. Edited to first check that the
update is in a terminal state and then verify that the CommitQuerier
makes its fidl calls to be more resilient to the speed at which the
test runs.

Original-Fixed: 91985

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650353
Original-Revision: cb7c706aa0c53b4dabaae78831c85e88dc42f8a9
GitOrigin-RevId: b2a8b7c59cee96363a72d42e43c4a0fc8cd340e5
Change-Id: I8173efe4c9a083d84435cc16607ca92db21cf858