[roll] Roll fuchsia [intl] Replace the use of sr-RS with en-US in an example
The sr-RS locale is still evolving, making change detector tests like
the ones edited in this change brittle when ICU versions are being
updated.

Change detector tests aren't the best approach for L10N-ized tests,
but in this case the impact is small and infrequent such that it
doesn't really seem worthwhile to change.

We really just want to verify that the language is changed when
the locale changes and that is it.  For that reason use locales
with more mature data in the test, such as en-US.

This should also allow us to update the ICU library use in
Fuchsia from 69.1 to 70.1.

Original-Bug: 92408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652254
Original-Revision: d569d88734f30ce3d987e3c80005fc3f3f6e2227
GitOrigin-RevId: 1f6bed2e3862190e22e5c1bcaed11629f67e2760
Change-Id: I2b862a31600c4ae007373b5ffe556d700a784401