[roll] Roll infra/recipes [fuchsia] Copy build logs from subbuild
The build recipe module attaches many useful logs to the top-level
"build" step, but for builders that use a builder-tester split you have
to click through to the subbuild to view those logs. To get around that,
we now copy through the logs to the parent "build" step, which should
make things more ergonomic.

It will significantly increase the size of the payload returned by `bb
collect`, which will now include all the steps from the child build, but
I think the tradeoff is worthwhile due to the usability improvement.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/652062
Original-Revision: 410761184eb7c39f3ea97c06e8c70754d337bb41
GitOrigin-RevId: 93af674e3be4d234ff60fb5a5ade981a7fde6a50
Change-Id: I65dd0b52fe35bde7bd953571c15a913bf871cfcf