[roll] Roll fuchsia [rust-3p] Actually preserve line endings
According to the [git docs], `text=false` is not a valid setting. This
results in the setting being treated as unspecified, which means git
will use the setting for `core.autocrlf`.

This fixes this, by instead using `-text`, which informs git to not
attempt any end-of-line changes.

In addition, this corrects the vendored gpt crate, which is using crlf
line endings.

[git docs]: https://git-scm.com/docs/gitattributes#_checking_out_and_checking_in

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652622
Original-Revision: 94f441c4c50c830653620f96e6e59c5c7b664fdf
GitOrigin-RevId: 4e8187b8875dd940c23fa61c5e11d8fd18e9fffd
Change-Id: I9d5b2a75cfbe3bf95a5b2c8c30479b1becc26cad