[roll] Roll fuchsia [zircon][sdk] Fix UB when calling `open`
Add missing parameter to `open` calls. The `open` function is a variadic
function which requires a third mode parameter to be set if O_CREAT is
specified [1]. This leads to UB when O_CREAT is specified without the
third mode parameter, as it interprets the following bytes on the stack
as the mode bits [2][3]. The same issue applies to `openat`.

Filesystem permissions rely on connection permissions instead of POSIX
permission bits on Fuchsia, and only use the mode type bits. Thus in
the worst case, the `open` (or `openat`) call may fail spuriously if the
interpreted mode bits are inconsistent with the flags argument, but will
not lead to any permission escalations.

[1]: https://pubs.opengroup.org/onlinepubs/007904875/functions/open.html
[2]: https://cs.opensource.google/fuchsia/fuchsia/+/80afb44f:sdk/lib/fdio/unistd.cc;l=1096
[3]: https://man7.org/linux/man-pages/man2/open.2.html

Original-Bug: 94471
Test: No functional change. Filesystem permissions on Fuchsia do not use
the mode bits, and they are only used to validate flags. Existing tests
that cover usage of `open`/`openat` have been updated accordingly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651611
Original-Revision: 7c0eb94ee99fae628fa4fc66f8a7c3c0ea96a100
GitOrigin-RevId: 6c46dfaab0844b5009c4344b111959146e79622b
Change-Id: Icdd57e05f66227d56529d573a8e9fc2b3027373c