[roll] Roll fuchsia [storage] Fix UB when calling `open`
Add missing parameter to `open` calls. The `open` function is a variadic
function which requires a third mode parameter to be set if O_CREAT is
specified [1]. This leads to UB when O_CREAT is specified without the
third mode parameter, as it interprets the following bytes on the stack
as the mode bits [2][3]. The same issue applies to `openat`.

Filesystem permissions rely on connection permissions instead of POSIX
permission bits on Fuchsia, and only use the mode type bits. Thus in
the worst case, the `open` (or `openat`) call may fail spuriously if the
interpreted mode bits are inconsistent with the flags argument, but will
not lead to any permission escalations.

[1]: https://pubs.opengroup.org/onlinepubs/007904875/functions/open.html
[2]: https://cs.opensource.google/fuchsia/fuchsia/+/80afb44f:sdk/lib/fdio/unistd.cc;l=1096
[3]: https://man7.org/linux/man-pages/man2/open.2.html

Original-Bug: 94471
Test: No functional change. Filesystem permissions on Fuchsia do not use
the mode bits, and they are only used to validate flags. Existing tests
that cover usage of `open`/`openat` have been updated accordingly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652103
Original-Revision: 145500a75f265a0c4ad61121d620e9f04a4adcbf
GitOrigin-RevId: af4b5512ade8af6c6fdf713a7be23132d2828f8a
Change-Id: I206d4476120f236cde739bc715169fc2ba90617b