[roll] Roll fuchsia [fidl] Natural encoder validation interface
This introduces a mechanism to perform valdation in the natural type
encoding process. Validation is only implemented for bools for now, but
is intended to extend to all types. This CL is mostly being sent out to
confirm the direction to go in and to create a basis for adding further
validation checks.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650941
Original-Revision: e8ea557314a6d11bfb20283c9ac2f50b7f580a3e
GitOrigin-RevId: 79b31df097890b51814dba7ee2cd190cd11a35d3
Change-Id: I853395503f1efd71f414e2906e878a87abe90078