[roll] Roll fuchsia [fuchsia-zircon-tests] Port fuchsia-zircon-tests to V2.
This change ports fuchsia-zircon-tests to a V2 test. We are attempting to remove
all instances of fuchsia.boot.RootResource from appmgr so we no longer
need to route the sensitive capability down that far into the system.
Note even though only tests use this capability due to the legacy of appmgr
we still have to route the capability down in production so we would
like to clean that up correctly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648388
Original-Revision: 06e7bcc7740f4e16020ee9e45b1cd59fe0f822a9
GitOrigin-RevId: 1da7c183601ab1fab1b9a006b63a69daa4124ecc
Change-Id: I21bcca8188c1e146886b099ebd5dc39c4d0f434a