[roll] Roll infra/recipes [fuchsia] Stop explicitly copying fint params log
As of fxrev.dev/652062 all build logs are automatically copied from the
subbuild to the parent build, so we no longer need to explicitly pass
the fint params to the parent build via output properties in order to
log them there.

It's also good to not include the full fint params text in an output
property, since people tend to want to use that in scripts and
post-processing jobs even though it's not a stable format.

This is not safe to land until the one consumer of the `fint_params`
output property has migrated off of it (http://cl/432034389).

Original-Bug: 94627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/652063
Original-Revision: 3cc51e0b4e5f4a18a2fe240885d4e498aac9b223
GitOrigin-RevId: 0fe693eb57c69838dcb6106291f30c8486402243
Change-Id: Ia4880eaf7db258dc7dad92d3d2d277866b86c132