[roll] Roll fuchsia [vm-stress-test] Port vm-stress-test to V2.
This change ports vm-stress-test to a V2 test. We are attempting to remove
all instances of fuchsia.boot.RootResource from appmgr so we no longer
need to route the sensitive capability down that far into the system. Note
even though only tests use this capability due to the legacy of appmgr we
still have to route the capability down in production so we would like to
clean that up.

Testing:
I tested the change on a qemu-core.x64 and it started up and ran correctly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648382
Original-Revision: ea5e0524f8e86e8f7df6bf528ef6fc6cf6ec1048
GitOrigin-RevId: 83682666c24feffcde73881242fcf96dcdca92fd
Change-Id: I0a16aa4d41ae456a08c6f3bf29c4a27703536be7