[roll] Roll fuchsia [virtualization] Add batch operations to BlockDispatcher
This change extends the BlockDispatcher interface to provide operations
to read/write batches of requests and the virtio_block device is updated
to submit all the request in a single chain as a single batch.

None of the dispatchers are updated to provide custom batch operations
yet. Instead we just add a default implementation of ReadBatch and
WriteBatch that call the existing ReadAt and WriteAt methods in a loop.

Original-Bug: 93355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652745
Original-Revision: b140d82dd3ac80e23d728dc24a014eb9fb82558d
GitOrigin-RevId: fdbb6e9708ee4b00ab34c468d6b9dfbc7eab67f2
Change-Id: Id2466488cfe78f6b39494beea74f505062c8301a