[roll] Roll fuchsia [fidl][cpp] Design doc for unified bindings
Since we would be referencing sections of the design doc in CLs quite a
bit, this CL makes the design doc public, and applies some wordsmithing
and editorial adjustments to make it easier to read.

In doing so, some obsolete content were removed. Some more rationale was
given behind particular decisions.

As of now the doc is not written up to "the RFC standard"; its role is
closer to a living reference that we could point to in public, and it is
expected that parts would be amended as we learn more through the
iteration and user feedback of the bindings.

Test: checked for style violations via
    fx mdlint --enable all --root-dir src/lib/fidl/cpp/docs
Test: staged the doc on devsite

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614030
Original-Revision: 1ca447a03bba1fb2b8bcbfe3bf7e58104ce490e4
GitOrigin-RevId: 3188564f05e19cab94b9b10de7a2ff154d9c0f3b
Change-Id: I2a36a13ce898f0f86834a338fc1c62733cae6aa8