[roll] Roll fuchsia [storage] move fs services out of svc
When filesystem are launched as processes, they populate an outgoing
directory with the filesystem root and a variety of services.
Originally, the services would be put in a /svc directory, to mimic how
they are constructed for components.

However, this isn't what the exported directory looks like when a
filesystem is actually launched as a component. Instead, exported
capabilities are put in a single, flat directory. Filesystems should
mostly look the same whether launched as a component or as a process, so
we move the filesystem services to the root of the outgoing directory
when launched as a process.

This makes it simpler to write tooling that can interact with a
filesystem instance agnostic to how it was initially launched.

Original-Bug: 91577
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652323
Original-Revision: ba7de0910c73a0780b2835017e60719aa75c23ab
GitOrigin-RevId: a8d595769d60d1c113c0c0a230ea0fa7ee550be9
Change-Id: I1cb8ffae1228ee474f85cef67deec44cf67de1a5