[roll] Roll fuchsia [media] Return status from call failure
A recent CL
Ic56ed1521003f3f382549e157158be0dfcf1fdd3
added an assertion on call failure, which may not be safe because it
crashes the process on failure. This CL updates the behavior to return
the status.

Let me know whether or not this is the correct way to handle a call
failure.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653922
Original-Revision: 4418d67cc2a678cfa6a2af0fff27c7a723a61fa4
GitOrigin-RevId: b642ea4b27675ec8dd64e578586ea361faa798b9
Change-Id: Ia54f5467714334a6b1a66da78ef2681a0d45f5a2