[roll] Roll fuchsia [fidl_driver] Tests for client and server reset
This CL adds tests that the correct callback is invoked by the server
when the client end is reset and likewise by the client when the server
end is reset.

This only covers the WireSharedClient case, but the WireClient case
can be added once WireClient calls are working properly. It is
arguable that it may not be needed, though, since the logic largely
shares the same code paths.

Multiply: fidl_driver_transport_tests (fuchsia): 100

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650401
Original-Revision: 560ff3516de89050be30286fc4603e05040ac843
GitOrigin-RevId: 92299b137ee18d6ed4509c66ddb7162d6b2cbcdc
Change-Id: Ia5eae4c54d600706dc729d88bead1a603d7c4163