[roll] Roll fuchsia [flatland] Rectangle Compositor turn color conversion on/off
Create two render paths in the RectangleCompositor for rendering
with and without color conversion. Having a renderpath that does
not use color conversion will be useful for when there is no color
conversion data turned on, and for when for future group opacity
related usecases where intermediary passes do not require and conversion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650505
Original-Revision: 9ebcf2d80fbe0d757528ba4116429f9a3f114951
GitOrigin-RevId: a77ab3cca19e9ccd4e2d9579e60b99b1a799295d
Change-Id: Idc1c02e1f7394ff2870f3bb8b61161842c249600