[roll] Roll fuchsia [netstack3] Route locally destined packets to loopback
This change updates one hack and replaces it with another. The original
hack looped-back locally destined packets at the IP layer. This change
updates that hack to instead route locally destined packets to the
loopback interface and then allow the stack to accept any packet
arriving at the loopback interface that is destined to an address
the stack owns.

Although we still have hacks with locally destined packet routing,
this change moves us one step closer to end goal of having all
locally destined packets flow through loopback but with routing
rules. See https://fxbug.dev/93870 for more details.

This change also moves the hack of using the loopback device to loopback
locally-destined packets outside of the socket layer. Instead, the socket
layer simply has a context trait method to return a route which performs
the hack. This also allows the IpSocketContext trait to be trimmed to
its only dependency on the rest of the stack - route lookup.

Original-Bug: 93870

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647340
Original-Revision: 7116912968983fc95a2bdf7939b08186036ff660
GitOrigin-RevId: 6fae88ffad8e419b5bc0df936b01c9132583c516
Change-Id: I74ab9f97ee20abf9fbabbd2166210490c9222fa5