[roll] Roll fuchsia [netemul] Serve test protocol until component shutdown
Currently, the netemul_test_runner bails as soon as it sees an invalid
configuration in a test component, closing both the fuchsia.test/Suite
channel (on which it communicates with the test_manager) and the
fuchsia.component.runner/ComponentController channel (on which it
communicates with component_manager). This leads to a race between
component shutdown and test failure.

To synchronize, continue to serve fuchsia.test/Suite (closing incoming
request channels) until the runner receives a Stop request from the
component_manager.

Also, add integration tests for invalid test configurations, but include
them in a separate package so that we can allow ERROR logs in these
tests, because tests with invalid configurations cause the
netemul_test_runner to emit ERROR logs.

Original-Fixed: 93307
Multiply: netemul-runner-tests
Multiply: netemul-runner-errors-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651617
Original-Revision: 76431daed912c00d4fbd8363bbcfb645c540feb9
GitOrigin-RevId: 0fdff391f243df65f409431a3dc5f4ddd8e0266c
Change-Id: I310c3dd477d7a7a1fbec562ce1f23a1fe0177d4a