[roll] Roll cobalt-registry [internal_metrics] Add replacement metrics for component-based metrics
Three of the metrics in the cobalt internal metrics project uses
component strings, which is not supported in Cobalt 1.1.  Add
replacement metrics for these metrics, using a dimension instead of a
component string.

The two metrics that track bytes stored and uploaded will use an INTEGER
metric type, while the metric that tracks logger calls will use an
OCCURRENCE metric type.

The dimension will enumerate the list of Cobalt projects, and will need
to be periodically updated as new projects are added. The client code
will also need to maintain a mapping of customer/project IDs to these
enum values in order to log the correct dimension value. This was chosen
over using an un-enumerated dimension with a max event code and logging
the customer and project IDs directly, due to limitations on the max
event code value being lower than the legacy project IDs, as well as
that pattern being a non-ideal workaround not suggested for typical
uses.  Additionally, the mapping of projects should be fairly small so
should not provide significant impact on the client's size and memory
usage.

Original-Bug: 64989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/652926
Privacy-Approval: Steve Fung <stevefung@google.com>
Original-Revision: 915d9dae578491f22652307157fe7ddb019648ca
GitOrigin-RevId: d0e0e7063d3c2c53581d5bf5697434a0cc2d1368
Change-Id: Ia3c42e1ffe462f365fd26c46c48b477678fa0966