[roll] Roll fuchsia [component][fuzzing] Add size constructor to Input
This CL adds a constructor that allocates an Input for a given size and
value-initializes it. This is especially useful when constructing inputs
in the capture-list of promises that will fill the input in
asynchronously.

Bug 92490

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640204
Original-Revision: a6f7c8b8c477a0c35b300d4cc4a0a0de7d0cfc20
GitOrigin-RevId: 228ab805d2823986190a0699d55f50fb3dffc7ab
Change-Id: Ib5db83c3a3eebbb9c0d4ae8c1e6ccdad7d17e42c