[roll] Roll fuchsia [vfs] Use non-discardable fidl::Result
fidl::Result will soon be marked [[nodiscard]], making it necessary to
use the value. For the sake of this CL, the result is cast to void,
keeping the existing behavior.

Note: the first version of this CL checked that the result's status
was ZX_OK, but this failed on CQ. It may be useful to follow up on
this CL and diagnose the reason for it.

Original-Bug: 94739
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653429
Original-Revision: 365ff6a582c14778704f5e244d731dd06281f83c
GitOrigin-RevId: 2017cc762ca600d3bb1c759fb88210ca1be4728c
Change-Id: I9866a58911fb915b09e8c410209a4688035aa0a7