[roll] Roll fuchsia [devices] Use non-discardable fidl::Result
fidl::Result will soon be marked [[nodiscard]], making it necessary to
use the value. For the sake of this CL, the result is cast to void,
keeping the existing behavior.

Note: the first version of this CL checked that the result's status
was ZX_OK, but this failed on CQ. It may be useful to follow up on
this CL and diagnose the reason for it.

Original-Bug: 94739
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653016
Original-Revision: 202ca1751d934cd14a0945517e610954ac5ac95b
GitOrigin-RevId: 70a78849c14c6d9dda1a264a4cbcabbaa885995e
Change-Id: Ie3de087e59ece0c84e6114432319e8cbae7361b6