[roll] Roll fuchsia [build] Replace rustflags target-feature=+crypto with +aes,+sha2
Something changed recently in rustc that causes it to generate warnings
when building for arm64 targets.

This appears related to https://github.com/rust-lang/rust/pull/87729
(though that PR is dated), which indicates that rustc is removing the
crypto target and replacing it with "aes or sha2".

This PR adds both, and removes crypto, for Rust only. (This does not
work with clang (cflags).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654243
Original-Revision: 7309bf5903dc36421531f34d317e95509b71210e
GitOrigin-RevId: 6403591387013e6af47658556cc2b5096d42b28b
Change-Id: Id9ba8b7c61f9376f93a12895e4427ddb2675d300