[roll] Roll fuchsia [fdf] Block drivers from driver index
Driver index will not return a driver in MatchDriver and MatchDriverV1
FIDL response if the driver is marked as not allowed. Driver index
retrieves the value of the boot argument `devmgr.disabled-drivers` to
determine which drivers are not allowed. The value of
`devmgr.disabled-drivers` is a string that is a comma-separated list of
component URL's of drivers that are not allowed.

Driver test realm will create the `devmgr.disabled-drivers` boot
argument using the realm arguments passed into it on creation.

Original-Bug: 92329
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644842
Original-Revision: ad617e605008b6ef340c9cd4f162dabb8ff2d71f
GitOrigin-RevId: 6b030b55199b5dd19d310f133cad6ad364572162
Change-Id: I6f017b0925e58785d54e643c191437af5e0dbb9e