[roll] Roll fuchsia [fidl][cpp] Fix llcpp_interop_tests flake
The test makes a call with an infinitely old timeout. As a result, when
the request was invalid, it will not immediately fail: sometimes the
channel call returns early enough with the expected timeout error,
other times the server reads the invalid message first and closes the
channel.

Also adds an assertion to make sure any message we `Write` or `Call`
down a transport is a transactional message.

Original-Fixed: 94400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654303
Original-Revision: f6907815fbf6dc7540e9683996f1e102d13b5944
GitOrigin-RevId: 598ec7b16c384d2443a026b6396901251582b20c
Change-Id: I9b5f81ba3a860a35fa66eaf683acc2cf88223f46