[roll] Roll fuchsia [bundles] Add a bundle for CTS tests.
The CTS CI/CQ builder will be made to use this bundle in a
later change.

This is needed for scenic CTS tests which depend on
scenic and hdcp implementations that are not part of the
core image.

Original-Bug: 94597
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651620
Original-Revision: 185dd7731daad04d586f8b6d5238a70f4f0ca18a
GitOrigin-RevId: e9cca6fd08e542ad885aabe13c2de4e054ca36c5
Change-Id: I338dbaaa393ee9b149d3b98b63063d745ac7c47e