[roll] Roll fuchsia [component][fuzzing] Add CorpusReaderClient
This CL adds an AsyncFidlClient class, and extends it into a
CorpusReaderClient. Additional test fixtures are also included to make
it easier to fake FIDL servers in async tests.

It does not add the CorpusReaderClient to the controller; that change is
saved for a subsequent CL.

Original-Bug: 92490
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642092
Original-Revision: 90c1c1fca9915994cb0557091213b2cae0a63c10
GitOrigin-RevId: 132566064f62516fcaf63cc7d658df86631e694e
Change-Id: Id4b5310549f08acb8c180aca9ad129d2f5e94dd9