[roll] Roll fuchsia [test_manager] Allow dart/flutter runners and route required services
In order to run a CFv2-based Dart test, I developed a Rust realm_builder
shim that loads the required runner into the dart component's
environment, and eagerly launches the dart test component.

Because the Dart-based test requires loading a Dart or Flutter Runner
into the environment (and the Runner requires some additional privileges
and services not offered to hermetic tests) the test must run as a
`system-test` at this time.

Also note that it may not (yet) be possible for the test manager to
capture the exit status from the Dart test to the test manager. The exit
status is typically used to properly report tests passing vs. tests
failing. However, with the described Rust realm_builder workaround, test
manager does report the test failure if the test logs an error (which
the dart::test implementation currently does).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652767
Original-Revision: 3c343f6c2829efbaacc2dfcb6a3f7020957ccc24
GitOrigin-RevId: cc301e3abf06933d865d67327056705fbd6322ed
Change-Id: If85f59af7b1f8a6088871260fc40b6ea19a16935