[roll] Roll fuchsia [component][fuzzing] Use CorpusReaderClient
This CL modifies the Controller to use the CorpusReaderClient.

It makes minor changes to the Controller unit tests, but does not
replace the FakeCorpusReader class in this change. This helps to
demonstrate the equivalence of the code before and after the change.

Original-Bug: 92490
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642093
Original-Revision: 7a0de1a39510253fe593360a965568ca525160c4
GitOrigin-RevId: 451ddefea371b96b05d128f04c1e47e66f03209d
Change-Id: Ia1297ee0f2bc6fde9bb7809cf0b0f7910cfc2cf3