[roll] Roll fuchsia [component][fuzzing] Add RunnerPtr
Shared pointers are easier and safer to use with asynchronous futures
and promises. This CL aliases std::shared_ptr<Runner> as RunnerPtr, and
updates call-sites to use it.

It also provide MakePtr as a way to create a RunnerPtr. Although not
needed in this CL, the Runner will soon want to access to the shared
ExecutorPtr. This CL includes the pointer as a constructor parameter
rather than updating all the Runner constructors again in a subsequent
CL.

Original-Bug: 92490
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644534
Original-Revision: 54b10e188bc7b26b21e7ceb2a7d2193cd7d4b68b
GitOrigin-RevId: a030fead0b7547e535bbffd0024b096a7e81893f
Change-Id: Ida98c94a897076d669817d9feaaf10cbeae761ce