[roll] Roll fuchsia [size-check] Run size_check as part of the build
The legacy tool and the new tool are both running
as part of the build and the output are compared
with a tolerance.

Size checker and budget generation are moced to
build/images/assembly to break the dependency cycle.

check_size_depfile tool enumerates the manifests
files used by the size_check so that it can be used
as depfile.

Original-Bug: 94824
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625562
Original-Revision: ee0f845965dad957c571bc1706fb8ffd02f64793
GitOrigin-RevId: 0c3b0813dbe9ddc254e164f9672e364f8efaefa9
Change-Id: I010f38d9af757bc93104152c0043def244bbf916