[roll] Roll fuchsia [storage][fdf] Provide read/write replacement
Provides a drop-in replacement for calling read
or write on a file descriptor.  The file descriptor
must refer to a fuchsia_io Node, which must be backed
by a block device.

This helper function will be used to replace the instances
of read and write in the storage stack.

Test: A test has been added in the fvm_integration_test.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648110
Original-Revision: 866ca02a9da001260f50fd65b24843f38d384606
GitOrigin-RevId: e9f438e430cbbb0b0ccd24633c7cd28d60c7a12c
Change-Id: I3648cb446a7df2a2b8be39258a409c8211b28480