[roll] Roll fuchsia [ffx][driver][dump] Display a separate node in the graph for each device with the same name
With the current implementation, devices with the same name are displayed as the same node in the resulting graph which is a bit confusing (eg. it looks like there is only one "hid-device" device with many parents.)

This implementation also makes it easier to customize each node with extra styles/info based on device info in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655402
Original-Revision: 118fad656f258112c56f41560eee2f065c94def3
GitOrigin-RevId: cc5de13f6753e8d3f1a00f1cb56a7593a0f91522
Change-Id: I150fb136dda19aeef8e21ca9e3a2b05bfa92a85f