[roll] Roll fuchsia [microbenchmarks] Disable building the host version of the tests for Mac
Disable building fuchsia_microbenchmarks for Mac, because debugging
problems with running this on Mac has turned out to be a pain.

For example, mutex.cc depends on the C11 <threads.h>, which is not
provided by the Mac toolchain.  (It's not provided by the
currently-pinned Linux sysroot either, but it should be provided in
newer versions.)

This relands part of this change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/647584

Original-Bug: 93071
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655665
Original-Revision: 4b0f9b30d29a3219278c81e170c0832da66fd008
GitOrigin-RevId: 390525001b09b0b666556e406487477722cac357
Change-Id: If7a931d6de71876ca6892a7ce367d33a6344a319