[roll] Roll fuchsia [hypervisor][apic] Support NMI delivery mode
When a guest writes to the ICR register in the x2apic with an NMI
delivery mode (0x4) then we need to generate an IPI to the target vcpus.
For these interrupts the vector field is ignored because the NMI vector
is already well defined (2).

Original-Bug: 93009
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655646
Original-Revision: 20129aee32e8bdd7d6df332b15990b4e8eec8259
GitOrigin-RevId: 69d329a5f0a3c68cade497182e54dfe2d3df2ebd
Change-Id: I183f34ebe75a57cf34147e5a55223e23877b658a