[roll] Roll fuchsia [netstack3][ndp] Verify address conflict in test
Change the test to first figure out what temporary SLAAC address would
be generated, then assign it before it is generated. Add a counter that
gets incremented on conflict and is read in the test. Together, this
verifies that the test is actually causing a conflict, before checking
the behavior that should follow after.

Without this, it's possible for the code to be changed to not handle
conflicts correctly and for the test to still pass.

Original-Bug: 21428

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652042
Original-Revision: 013892255e91e4b5684eb588749f327f8f355edf
GitOrigin-RevId: d1acdaa59a5717732770145e7e29621bdbabc7ee
Change-Id: Id15f6a9e75e581d038857ba7a5e89c92d96691e8