[roll] Roll fuchsia [build][rbe] Suppress --canonicalize_working_dir.
For this target, the Rust command depends on the output-dir
as-is, and cannot be relocated to a different output-dir.
Mark it as such with the pseudoflag:
--remote-flag=--canonicalize_working_dir=false.

RBE accounts for this in the action hash, and will do the right thing
w.r.t. caching.

Original-Bug: 94355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655324
Original-Revision: 71067554c97fa45115bd293e183d9aca3805239b
GitOrigin-RevId: cb0230ac9083475d638798420ec3901ade70c5ed
Change-Id: Ieb806d0637fd69e25bdf030c2e0dce51f37c997d