[roll] Roll fuchsia [fidl][cpp] Introduce |AnyErrorIn<Method>|.
Adds the |fidl::AnyErrorIn<Method>| discussed in design doc [1].

It's primary purpose is to fold the transport error and application
error type of a method into one object, and to print a description for
either the transport or the application error.

To accomplish the latter, we introduce a `fidl::internal::DisplayError`
trait, which is implemented for int32, uint32, and all 4 byte enums,
covering all types that may be used for errors, in addition to
`fidl::Error`.

`DisplayError<T>` is used by `AnyErrorIn<Method>` as part of formatting
description.

Since logging is a common occurrence, we use two techniques to reduce
binary size:

- Use snprintf similar to |fidl::Result|.
- Use a |AnyErrorInBase| base to factor out non-templated logic.

[1]:
https://fuchsia.googlesource.com/fuchsia/+/refs/heads/main/src/lib/fidl/cpp/docs/structure-of-unified-cpp-fidl-bindings.md#natural-api_asynchronous

Original-Bug: 90111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655145
Original-Revision: 740851af097b5a7f62c742dee24e2f5b47fda37a
GitOrigin-RevId: f0c7d772f773fa06f11972c849831a83080c2acc
Change-Id: I0818b7e3f5ae9dee4e7efd7c149bd529892a0f39