[roll] Roll fuchsia [assembly] Tool providers can use paths to access tools.
With this change the zbi signing script can now be a tool, because the
tool providers can create tools by their path. This is important so that
we can log the usage of the zbi signing scripts.

Original-Bug: 92883
Test: fx test assembly_tool
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654264
Original-Revision: a44af503b9dc50c6cb0bd1ec0b676d9fdd315917
GitOrigin-RevId: 14519cd4b789837a58895473be3faba30dc0a269
Change-Id: If875cdde373693e47a8e55d3e2d3226608f3085e