[roll] Roll fuchsia [assembly] Use the tool provider for zbi signing
This change adds the zbi signing script to the command log by running
the script via the ToolProvider interface.

We are doing this so that we can ensure that we do not break zbi signing
when we switch to create-system.

Original-Bug: 92883
Test: fx test assembly
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654267
Original-Revision: b9d47a6dbc851a387e1dc3a336213c965369eb32
GitOrigin-RevId: 584c6b9d4d55c747a885e4fe2a25704eb6061359
Change-Id: Icf696d41a8319ea28cdf23111fc48dcc86bacbcb