[roll] Roll fuchsia [fms] add device specific configuration to virtual device spec
This is a reland of fxr/646802 with a new check for whether or not
we should generate a physical or virtual device spec or both.

From fxr/646802, this commit allows the default window-size to be
overridden using a gni variable. In order to distinguish between
similar virtual device specs using the same board but different
window-size, adding a prefix var is required as well.

Original-Bug: 88619
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651724
Original-Revision: 2b51cf7b832872082695ce973c8ae7580417b778
GitOrigin-RevId: 9948ca3ffcc5e6a7c5f30e78c4dade51e4ceb944
Change-Id: I4ee89672f558aabfea860a697c886afeac979246