[roll] Roll fuchsia [assembly] Strip both prefixes when comparing the command logs
When generating the zbi, a path to the kernel is provided, which is in
the legacy bundle, which uses the reference outdir. This means that when
stripping the paths for the comparison we need to strip both outdirs.

This is an example of the path used in both command logs:
  obj/build/images/zedboot/zedboot/legacy/kernel/kernel.zbi

Original-Bug: 92883
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654274
Original-Revision: 5709fa1dc7fa2e946ab71d74b22211c1dd147294
GitOrigin-RevId: 8317dba9334c7e399132d7cf52e98282e96af124
Change-Id: I1b5aa769a634969e6ed85ec0c8802ff656188621