[roll] Roll fuchsia [lib][zircon] Remove test for debug address
This feature is tested extensively in utest/debugger, and the getter
and setter implementations are done by a standard macro. The issue with
testing this is that the debug address of the process may be set by
other parts of the system, and the property is only allowed to be set
once.

Original-Fixed: 94568

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656003
Original-Revision: 3f7ff51d5aeddd4ee9833bb3c6c23937e7cef543
GitOrigin-RevId: 15f8caed3379953041995edb045ed90a187bd5b8
Change-Id: I811e9590486514776eacf9a63e7fd6fdb02310ed