[roll] Roll fuchsia [bt] Remove RfcommTest routes and define testonly manifests
Remove capability routes to `RfcommTest` so that it is not available in
production environments.

Instead, define testonly manifests with the capability routes. Users
will need to manually update core.cml to point to the testonly manifest
if needed. At this time, we do not want to go through the process of
making #bluetooth-core product variable.

Remove appmgr route to RfcommTest as it will be routed through the
testonly manifest.

Define a testonly group for Bluetooth core. This can be included in the
relevant product.gni file to set up a special BT realm for testing
scenarios.

Original-Bug: 86390
Test: fx build && manually verify bluetooth-core starts up okay
Test: fx build with bt-init-testonly.cml and verify RfcommTest is
available

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653224
Original-Revision: f1baaa79706adb41b1e69bf4843eddb5187430f2
GitOrigin-RevId: da3de1a9f100aefc97a4949c878d6f265127184d
Change-Id: I7501de8825c759a2c1277b9cdcf92c2acdc00c0b