[roll] Roll infra/recipes [fuchsia] Add report_binary_sizes bool
The size checker is now integrated into the build. The build emits a
build API manifest containing a 0 or 1 length array of paths to size
checker JSON.

The intent of this flag is to:

1. Toggle reading of the build API
2. Assert that the array is length-1
3. Read the file and store the results as a cached property in
   build/api.py (replacing the codepath where we explicitly run size
   checker to get results).

This CL only adds the flag so that builders can set it before we
implement the cutover per ^ these steps.

Original-Bug: 95221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/657225
Original-Revision: 8e5da3f0ccd96c77bd7624c272fcde3463772546
GitOrigin-RevId: c6a9751c8fc5d6e4a3fc849b2f21986281ea4b88
Change-Id: Id40b7945b9afa0804ab83c08be9eefa18e177ca6