[roll] Roll fuchsia [ffx][repo] Fix ffx repository package {list,show}
These broke because we were prepending "blobs/" to the paths when trying
to determine the package timestamp. This was missed due to a lack of
tests.

Along the way, this changes these functions to use the target length
entry, rather than the custom field "size", which isn't used in the
package resolver.

Original-Fixed: 79915
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656663
Original-Revision: 2bfd43edd136ca2bcbfe2dac0f57eb24aa4b79a4
GitOrigin-RevId: ed87d4d8f8ec7ef126f15b935efd0f3fe9cd1999
Change-Id: I777eb842ae1419ed6963b806b291d21ce5d0eeb6