[roll] Roll fuchsia [fidl][fidlc] Add TopResponse shim
The `TopResponse` generated naming context should not be exposed to end
users of bindings. Further, the naming context should not be used to
describe type except for the fidlc-generated, top-level "wrapper" struct
created when users use the `error` syntax on their FIDL methods.

However, in the current pre-union/table payload world, all anonymous
struct payloads for methods not carrying the error syntax still receive
the suffix `TopResponse`, like so:

```
protocol MyProtocol {
  // Response struct is named "MyProtocolMyMethodTopResponse"
  MyMethod() -> (struct {a bool; b int32; });
};
```

This is fine for struct payloads: unified C++ bindings aside, we
"flatten" the struct into its constituent parameters, ensuring that end
users of the generated bindings never touch the `TopResponse`-suffixed
type. For example, the Go bindings generate the response signature
`MyMethod() (bool, int32, error)` rather than `MyMethod()
(MyProtocolMyMethodTopResponse, error)`, as would be expected in the
"unflattened" case.

Unlike for structs, the name of the payload-carrying type must be
exposed for union/table payloads used in this manner, as those payloads
are never flattened, leading to ex. the second Go signature from the
previous example. Instead, we would like the name of the unflattened
return value to be `MyProtocolMyMethodResponse`.

One option to achieve this would be do a source-breaking change, and
just have all naming contexts for payloads in this position (even
existing struct ones) resolve to `___Response` rather than
`___TopResponse`. This is not desirable, as there are already use cases
where this would cause a name collision (example:
https://cs.opensource.google/fuchsia/fuchsia/+/main:sdk/fidl/fuchsia.ui.observation.geometry/provider.fidl;l=77;drc=b9d4f2421059c5d01b66ce476887dd1554d8d235).

The other alternative is the one implemented by this CL: all non-error
syntaxed anonymous response payloads are now named `___Response` instead
of `___TopResponse`. However, in the case of only struct layouts, a shim
is later applied that reverts the names of only those payloads back to
`___TopResponse`. In the future, once `TopResponse` wrapping is removed
completed, this shimming function and the lone call into it can be
removed.

Test: fx regen-goldens; fx test
Original-Bug: 95231
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657462
Original-Revision: 40840e19e5307ca01a92b1e843323f369497a6d8
GitOrigin-RevId: 5e1f91df5d831083af5886273a55ebd95a234f31
Change-Id: I92c5d110953e8da5d9fb7362e2a51084869cff41