[roll] Roll fuchsia [driver_manager] Fix fuchsia.compat.LIBNAME property for DFv1
BindDevice() is called with an absolute path to a library, but the
fuchsia.compat.LIBNAME property should just be the library's name.

We fix this by adjusting driver_host to pass just the libname to
driver_manager, and then driver_manager uses that to match the device.

Original-Bug: 91608
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654884
Original-Revision: 03244e3124ac5599067f6e9cb7417d517b085333
GitOrigin-RevId: 233b9c0eeabff27ac5fb3a1d88d84ae7bef9adbf
Change-Id: Ib2bce36660fac3fa3fe65f48dc7f3e18e6ce66ae