[roll] Roll fuchsia [ffx][preflight] Gather CPU features from cpuinfo
On Discord, roddux#5128 (an external contributor) pointed out that
`lscpu` makes use of gettext, and that our regex assumes that the person
running `ffx platform preflight` is using a system configured with
`en_US` i18n/l10n settings.

This change switches from using output of the `lscpu` command to looking
for either the `vmx` (for VT-x) or `svm` (for AMD-V) flags in its
output. We may like to consider writing an API for consuming
`/proc/cpuinfo` output, but that's outside the scope of this change.

Test: `lscpu` tests modified to check cpuinfo from real-world devices.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658282
Original-Revision: ed83be4b9efbc17780d8b27fec056308b0b1a9fb
GitOrigin-RevId: 5f3fd2e131bb90d4218bf2399d760df2554f63ec
Change-Id: I43a87310ce2ec686fb9cf5db5d14a93fc93605a1