[roll] Roll fuchsia [kernel] Update spectre mitigation to reflect is_user_accessible changes
As of fxr/639991, is_user_address was replaced with is_user_accessible.
Rather than checking if addresses are within the range
[USER_ASPACE_BASE, USER_ASPACE_BASE + USER_ASPACE_SIZE), we instead do
an arch-specific check for "can this address be accessed by the user".
This conflicts with spectre mitigation which still checks that a user
address is instead in the range
[0, USER_ASPACE_BASE + USER_ASPACE_SIZE). This updates the spectre
mitigation to conform to the new logic behind user addresses. That is,
if there exists an address outside the range [addr, addr+len), then
squash `addr` and `len` to zero.

This will allow for fewer changes to be made when passing tagged
pointers to the user_copy functions.

Changes:
- Rename confine_user_address_range with validate_user_address_range,
  and update its implementation.
- Fix issues found by the test_addresses_outside_user_range test that
  resulted from conflicting logic between confine_user_address_range
  and is_user_accessible_range.
- Remove test_confine_user_address_range which explicity tested
  behavior in the old confine_user_address_range function. Instead,
  update arch_is_user_accessible_range to also check its various cases
  against validate_user_address_range. This makes it easier to reuse
  existing test cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655813
Original-Revision: 53bae38c2728e2f48bf110f4954fb331d4202cb7
GitOrigin-RevId: 7510c1f9ed06aeae7db92ef4ba32b582621659de
Change-Id: I023202e279ed0b13ad2d4551ba1d34fea97c660a