[roll] Roll fuchsia [audio][drivers][hda] Add MISCDBCGE bit clearing
It is recommended by HW that the CGCTL register MISCDBCGE bit is
deasserted when a reset of the HDA controller HW is performed.
This bit controls the "Clock Gating Control"
"Backbone Dynamic Clock Gating".

This was found investigating very rare boot HDA initialization
failures in fxrev.dev/84037, but it is not known to fix these.

Test: intel-hda-controller-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658377
Original-Revision: dd7fc4538ee5db767abb41bca0bf0ffb1c161ac4
GitOrigin-RevId: be4ce9ad8160117eba24751c127d426947a0f33d
Change-Id: I65e678018447c1eef5c78e46a43ed691d72b5135