[roll] Roll fuchsia [network] Introduce todo_unused macro
The macro serves as a code manifestation of the cultural agreement to
balance two needs that are often in opposition:
- keep CLs small
- avoid introducing code that is not used outside of tests

The macro enforces that a bug number be informed to track the technical
debt with urgency, and annotates statements with `cfg(test)` which
require revisitation organically when the code becomes load-bearing.

This does not address the unfortunate virality of `cfg(test)`, but we
take the advantages above to be sufficient.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657592
Original-Revision: f4c011301f617d335935c0749697952d24d6bc72
GitOrigin-RevId: 92a38e2dce1107154d5daf92a6dea7b667f3fe36
Change-Id: If0b387bb62d758ceca0d5f8cbe98081cd93ae654