[roll] Roll fuchsia [audio][test] Handle single-frequency tests
A frequency specified in ".single_frequency_to_test" should
ideally be in kReferenceFrequencies[] so we have FR and SiNAD
results for it. If it is not, display a warning but continue,
actually testing at the nearest frequency that IS in our set
(rounding down); we do this (rather than CHECK) to make single-
frequency spot-checking easier during ad-hoc investigations.

Other adjustments in how we display single-frequency results.

fx test audio-core-fidelity-test                        \
        audio-core-google-pipeline-fidelity-test-nelson

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658584
Original-Revision: ddb51ae47caefe8f0d449e7de933907687efc74e
GitOrigin-RevId: e56630c41a93c4413bb6c929fff96bc94b856d92
Change-Id: I69c8d59b86c4fdd32896f8993260b09bc5068bcd