[roll] Roll fuchsia [fint] Stop building system_snapshot explicitly
The infrastructure initially explicitly built the system snapshot so
that the size data it contained could be uploaded to GCS. However, no
code accesses the file via GCS and we instead use a different set of
size-tracking infrastructure, so we can stop uploading it from the
recipes (fxrev.dev/658882) at which point fint doesn't need to build it
anymore.

This change is safe to land before the corresponding recipe change
because the recipes will skip uploading the file if it doesn't exist on
disk.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658748
Original-Revision: b86ca78e05c483ab07c7522373f29de0beeaf956
GitOrigin-RevId: 6108bc0ab2e9cb5f229f29fe7cb4d2d213fc15b8
Change-Id: I426eb0b53a9eb54918b1f9d842ca654233abb078