[roll] Roll infra/recipes [rust][toolchain] Fix statically linking libc++
This change looks like it disables static linking, but it actually
doesn't. The `static-libstdcpp` flag determines whether the
`LLVM_STATIC_STDCPP` environment variable is set in
`bootstrap/compile.rs`. That variable is checked in
`rustc_llvm/build.rs` to determine whether the directory containing
`libstdc++.a` should be added to the library search paths. However, we
don't use `libstdc++` (the GNU implementation), we use `libc++` (the
LLVM implementation). There's a special case that we want to fall into
in `build.rs` after the check for `LLVM_STATIC_STDCPP`. If we specify
`stdlib=libc++` in `cxxflags` then it will instruct rustc to link
against `libc++` instead. So in order to get to this case, we need to
set `static-libstdcpp` to `false` and add `stdlib=libc++` to our
`cxxflags`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/658886
Original-Revision: d36a3b608399d38367aead9ae9ede46f65101e61
GitOrigin-RevId: ff85149f7c4ddfdab1a2452dd800a41310110f22
Change-Id: Ia6e3d5750d4d15eab0bd9d76dc56419dd246d7c8