[roll] Roll fuchsia [starnix][binder] Implement ref-counting commands
These commands allow a binder process to increment/decrement the
weak/strong reference count of a binder object in a remote process.

In the future, the binder driver will keep its own ref count of objects
and only send a command to a binder process for the first and last
reference.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656566
Original-Revision: 70a9716cce87530f634f89982a85dc00d17af246
GitOrigin-RevId: 4845cebcb1e84f51dcf47c094ddfe1c1346921ec
Change-Id: Ic8965e0c6feabe9f4a027044b95ccce0e4a0c5ee