[roll] Roll fuchsia [ffx] Use correct ffx config API
If a user has cleared the state/config directory on the local machine,
the ffx emu tools would crash with an error of "no configuration value
found". This error came from the ffx config API, but didn't tell the
user anything actionable.

This CL adjusts the launcher's usage of that API, such that we use
"get()" instead of "file()". The former returns the requested
information as a string, while the latter expects the information to
point to a filesystem object and also ensures that file/folder exists.
Now, if the directory is missing, the code can attempt to create it
before returning an error.

This CL also adds error message context to each of the affected
subcommands, so that any other errors from the API call can be more
easily debugged.

Original-Bug: fxbug.dev/95568
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658863
Original-Revision: f758289f6b1b580354f97acd20d75efeb060dcc0
GitOrigin-RevId: 60a90b8b19f768e5c956bdc0ea56cd2cb615e4d0
Change-Id: I8f241fd8cc1aac9c8230aef3e7d8836492b01eb5