[roll] Roll fuchsia [netemul] Remove InterfaceOptions.enable_dhcp
Remove the option to start a DHCP client on the interface being
installed in a netemul runner test. This functionality is mostly unused
in existing runner tests, and is only used in a test that relies on the
netemul guest functionality, which will likely not be fulfilled by the
runner in CFv2.

Test: fx test netemul-runner-unit-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659185
Original-Revision: cbc61e499d43ab113b1314b8ab6afa2f9772e194
GitOrigin-RevId: ec9477bdaea7fe642a56d9847794ea267460dcb7
Change-Id: I04fb25b9ee41dc5973775d298b6117168a2f4bae