[roll] Roll fuchsia [mixer_service] Add a sketch of mixer objects
This CL adds a sketch of two kinds of mixer objects, Thread and
PipelineStage. There are two implementations of Thread, one for the
"detached" thread and one for the mixer threads. See here for more
discussion:
//src/media/audio/mixer_service/docs/execution_model.md

This is just a sketch with not much functionality. Future CLs will flesh
out this directory.

Original-Bug: 87651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649810
Original-Revision: fd95467a3b4601b66e1f9ca667590aa9597943a2
GitOrigin-RevId: 8c55cacaea19fda932304a2668147e6a6624ce8c
Change-Id: Ife6fa36859873d27cb666c5b4559ff9a7472867c