[roll] Roll infra/recipes [build] Stop uploading package snapshot to GCS
This file does not seem to have any consumers based on some digging in
Code Search, so there's no reason to keep uploading it. This upload
logic was originally introduced in http://fxrev.dev/250050, as part of
the http://go/tq-artifact-index work, but that was never fully developed
and has been replaced by the current size-tracking infrastructure.

This also removes one of the last places where recipes hardcode the path
to a file produced by the build, something we're trying to avoid in
favor of relying on the build API files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/658882
Original-Revision: 112e4c6d610c2596c23960538709f23b4162c1ab
GitOrigin-RevId: 4c4ba8ea87fff8d0ae089526b77c6976c2856a1c
Change-Id: I0f11e8b922893096de2e56123fefbeb5dc9b7b03