[roll] Roll fuchsia [devices] Make ExposedInspector a class
This CL changes the `driver::ExposeInspect` function
to instead be a class. This was done because after changing
its usage from `service::OutgoingDirectory` to
`component::OutgoingDirectory`, another object (a vfs) associated
with the diagnostics directory needed to be kept alive.

For the convenience of clients of this function, it was better served
to wrap the internal objects in a class instead of exporting it
as a tuple<zx::Vmo, fs::ManagedVfs>.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658504
Original-Revision: 443e7892a991369b09dcf31640f19b7c7c28d7d6
GitOrigin-RevId: e97730713dbf9ba90cf6d5c678e02494bb64177c
Change-Id: I5a3ab12ab7b4617cf7be8e0ebb2a523b7611af13