[roll] Roll fuchsia [emu] Include QEMU prebuilt in the SDK
We've been expanding the feature set for emulation for some time, and
part of the new functionality is the ability to launch a more recent
version of QEMU than the branch they forked for AEMU. Currently, the
kernel team tests with 5.0, but they're actively upgrading to 6.0 for
base QEMU. AEMU is already available as a prebuilt in the SDK, but QEMU
is not.

This CL adds QEMU to the set of host tools, while importing the prebuilt
artifacts for SDK inclusion. This enables OOT users to run the QEMU5
emulator directly from the SDK as it's delivered, matching the
functionality for the AEMU backend.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657486
Original-Revision: 0336131104d8617226f28a8fc6384d3881ecf2ca
GitOrigin-RevId: bcee998a771ca681f080f47c29bf7dfc11f4bb4e
Change-Id: I5e8b53dcb5c17eff5c6e8587e182bb72800fbfd5