[roll] Roll fuchsia [fidl][cpp] Cancellable callbacks in Zircon IPC
Introduces `Then` and `ThenExactlyOnce` which can be chained after a
two-way call to attach a continuation.

`Then` passivates the callback if the client object is destroyed.

`ThenExactlyOnce` calls it exactly once.

The next steps are to migrate user code and also update the driver and
natural messaging APIs to use this mechanism.

Original-Bug: 94402

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650825
Original-Revision: 72ea48e1227fe2ea6b3600e773bca795682160a0
GitOrigin-RevId: d73532ae1173d61fd69a0d471ed4cb8dbf9781de
Change-Id: Ie49baa64ccc381e0c063cd2a0f248acb7f221be4