[roll] Roll fuchsia [fs_management] separate out options, add unit tests
Move all the options to their own header file, add the ability to
generate the cli options based on the option structs, and add unit tests
for that conversion to make sure it's right.

This adds a natural place for doing the conversion to the startup fidl
struct when we add component launching, as well as a way to test that
conversion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658507
Original-Revision: 4ed05bca640c30e1b396b74817548747149916f0
GitOrigin-RevId: f4503d48c32ab86f7719debb6faa5d1c25616957
Change-Id: I75407d16dd9bf87d114df73c6e98893a678e2844