[roll] Roll fuchsia [blobfs] Remove decompressor sandbox from unit tests.
This adds a tiny threaded version of the decompressor sandbox meant for
use within tests. This makes it so that we don't need to additionally
include the decompressor sandbox component for tests. It is easier to
add 2-3 lines of boilerplate than to juggle around the component
manifests for all the tests that need it outside of blobfs.

Original-Bug: 62177

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630284
Original-Revision: 2d651a14ce08f456c52ace24fb04f57f886ee1a1
GitOrigin-RevId: d2add6dcaae37e8226fb42ec4ff4149ccbcc7464
Change-Id: I473eeab942ca775689e6e563db4638849b7b0c51