[roll] Roll fuchsia [camera][gym] Add frame arrival deadline
For Thermal testing, it is important to fail the test run if there
are indications of catastrophic failures such as no frames arriving.

This CL adds a deadline for frame arrival such a new deadline is
established at the beginning of each stream and at each subsequent
frame. The tolerance for lateness is a constant 10 seconds for all
streams (regardless of the frame rate).

When this deadline is tripped (for any frame), camera-gym will
assert and crash out.

The thermal test rig will detect this failure and fail accordingly.

Original-Bug: b/219806158
Test: Run thermal test normally.
      Run thermal test with forced sensor reset (test should fail).
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656783
Original-Revision: 20bc945f677e9046314259070b3c9b46e2b78657
GitOrigin-RevId: 31f0160dc310511fc68c2abfcdeed2c7b7160547
Change-Id: I9a378358dbcde4f185fe324cd62c265df808b899