[roll] Roll fuchsia [size] Remove board config dep from the size checker
The board config is deprecated now. This CL migrates size checker off of
it.

The size checker needs to know both the blobfs layout and the blobfs
compression variables in order to construct blobfs images. Currently,
the compression state is always 'true', so hardcode that for now. The
blobfs layout is passed in as a CLI argument that defaults to "compact".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657583
Original-Revision: db7978b00e3fb47a683807894544ea78479b88bc
GitOrigin-RevId: c6bdf90a416a6c8601fc89cdfbc9e208139fabfa
Change-Id: Idea656c291e19dcb0b12e64a57a9a4c1da4d63ec