[roll] Roll fuchsia [ffx] Indicate when femu fails to launch
When something goes wrong with launching the emulator process, we want
to inform the user rather than keep them waiting. This CL adds a check
to the "is_running()" function which does a "no-wait" join, to clean up
the process if it's defunct, then checks if it's still there. We then
call the updated function while we're waiting for the emulator to be
"live", and then short-circuit the health check and return an error if
the process is no longer running.

Original-Bug: fxbug.dev/94211
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659884
Original-Revision: 243f99430629330c443f3f24eddb1ade95187b34
GitOrigin-RevId: 284025580bdbf9e221871f904f169fa1e078b3de
Change-Id: I38542935d3b77ab1365814370f4f1d1fca568491