[roll] Roll fuchsia [usb][xhci] Clarify driver error messages.
When the USB XHCI bus driver fails to load because of trouble with the
resources which were passed to it, be clear in the error messages
about which resources were the source of the problems.  In specific,
when the VMO used to access registers has problems, report this
(instead of saying nothing at all).  When the interrupt object passed
is missing or invalid, complain that there is something wrong with the
interrupt instead of saying "MapMmio error", which makes it sound like
something went wrong with an attempt to map the registers, not that
the interrupt object was invalid or missing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659890
Original-Revision: 53853c0cba44d7de30302b6b76ed5643453b6920
GitOrigin-RevId: bf1229fdc24dbe5914152c6ec4090ec696e2af41
Change-Id: I4b65d97719845de6b248ee50e4d66987a1f918c4