[roll] Roll cobalt-registry [connectivity_wlan] Reorganize YAML anchors
This commit creates a separate section in the config where we can define
templates. These templates are just placeholder metrics that we can
attach YAML anchors onto. Each template has two variants: (1) one with
just the default reports, and (2) another where additional compatible
reports can be defined and reused.

Soon we will start using these metrics for experiments, so the use case
for attaching additional reports to the default ones will become more
common. At the same time, we want to avoid increasing the burden to
reviewers with the proliferation of YAML anchors. The idea is that
someone can first review the templates to verify that the custom reports
are compatible (and thus the YAML anchor for the metric portion and
YAML anchor for the corresponding custom reports are compatible). Then
when reviewing the individual metrics, they can quickly verify the
ones that only use anchors that are compatible with each other.

Original-Bug: 92093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/658922
Privacy-Approval: Steve Fung <stevefung@google.com>
Original-Revision: fc03aa4b5fde554736759fd2ef98c45598b7a5ea
GitOrigin-RevId: bb987079aad4353408575cf998e263ab79f1ed81
Change-Id: I5fa325d8947d224c160b5c78d4c3b810fec043c9