[roll] Roll fuchsia [netstack3] Ctx type carries separate context
Modify the `Ctx` type to carry separate `EventDispatcher` and
`CoreContext` objects.

This is part of a larger change in which all context arguments in the
Core are being split in two in order to resolve some lifetime conflict
issues and to prepare for a deeper overhaul of the way state is accessed
and synchronized. See fxbug.dev/92605 for more details.

Original-Bug: 92605
Original-Fixed: 93859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640181
Original-Revision: a82c9d3285a3522b097a18f8a485f8bdedbd34fc
GitOrigin-RevId: 8a4306af7ac20e2711a0f269b35942e017f76f21
Change-Id: I16a69657f96d4a71e7a17da3acde724503b6c1d2