[roll] Roll fuchsia [audio][test] Copy hermetic test lib to audio_core/testing/integration
During transition to CFv2 we'll need to keep around an old copy of this
library (CFv1) along with the new copy. Hence we start by making a copy
of the old library. I've moved this to audio_core/testing/integration
since this is a support library for testing audio_core.

This is a pure copy with some minor BUILD.hn refactoring: nothing has
changed except for #include statements.

Original-Bug: 90408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659392
Original-Revision: 9265923cfe729380150524f5012b5b98a9ea9551
GitOrigin-RevId: ede7826f686e1767d43181083311b49647f7b6a2
Change-Id: I473f0e0db48eaada36dfb0dc0bf040eb1f3e134f