[roll] Roll fuchsia [ffx][repo] Disable server by default
This changes ffx repository to initially have the repository server
start in a disabled state. This means that by default when ffx starts
up, it will not try to start the server, even if there are locally
configured repositories in registrations.

Additionally, this also changes `ffx repository server stop` to mark the
server as disabled if we experience any errors when trying to stop the
server. This covers the case where the daemon gets stuck and is unable
to disable the server. With this fix, restarting the daemon should block
the server from starting.

However, if you want the old behavior (where the server is started when
the daemon is started), users should do either:

```
$ ffx repository server start
```

Or the underlying config setting is changed with:

```
$ ffx config set repository.server.enabled true
$ ffx doctor --restart-daemon
```

We'll switch back to not starting the server if the user does:

```
$ ffx repository server stop
```

Or changes the underlying config with:

```
$ ffx config set repository.server.enabled false
$ ffx doctor --restart-daemon
```

Original-Fixed: 95279
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656487
Original-Revision: 03735a2f811fc2b8a6a1154e0729d8c0f57d78f0
GitOrigin-RevId: aba2a98f09f8ef56c4537331d560df22a06f6e25
Change-Id: Ia5e52e66ac3e7c60ecd4f4e00125611e58ea8699