[roll] Roll fuchsia [size] Do not fail build with budget exceeded
We need to upload the gerrit size report to display the size analysis,
therefore we cannot cause the build step to fail by running the size
checker. When a budget is exceeded, we will print a warning to stdout,
and return exit code 0 (success).

Later, we can give infra an indication of whether the budgets were
exceeded or not.

Original-Bug: 95016
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660023
Original-Revision: b009d49dbbf00b6bc9b32db28c0a3063fe82e881
GitOrigin-RevId: 804d9d83072ff4fa16178b8e06d8171db320d502
Change-Id: I813a7c26f515eb7ad57bb5ee88cac1279d93ec15