[roll] Roll fuchsia [usb][xhci] fix device state locking and check for disconnect
Originally we were getting the transaction_lock for a certain device
state to check if it is currently disconnecting, releasing the lock then
re-acquiring the lock to do something else. In between releasing the
lock and re-acquiring the lock, the disconnecting state may have
changed. Here we modify the code so that the lock is held until the end
of the operation so that we can be sure that the the device is not
disconnecting.

This CL also adds checks to disconnecting for device states in some
places that were originally missing.

Test: tested on Atlas

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656575
Original-Revision: 8e6b36801b65e65eb6ea2e9111de904999a33e68
GitOrigin-RevId: e14654039f9a65a885133d87ed84d76deeaa6a9c
Change-Id: I7aaa6c3aa5bbf614a348ed0df8df16307a7d2501