[roll] Roll fuchsia [wlan] Use security authenticators in SME.
This change uses security authenticators to negotiate a `Protection` in
SME. This includes refactors to code that reads authentication
configuration and RSNE data. A `SecurityContext` has been introduced
that both implements RSNE and provides a conversion API for constructing
security authenticators (in various contexts). Note that this is a
preliminary change and should not alter the behavior of SME.

The important change is in `on_connect_command`, where a security
authenticator is constructed from an `Authentication` message. Today,
some temporary code constructs the `Authentication` using the same
heuristics already present in SME, but this will be removed in favor of
changing the SME `Connect` FIDL API to include an `Authentication`
(rather than a bare `Credential`) in `ConnectRequest`s. Once that change
is made, SME will only attempt to use the security protocol specified in
the `Authentication` and will no longer make choices regarding the
specific protocol.

This change was factored out of a larger abandoned change: fxr/643789.

Test: Refactored unit tests and manual testing.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659442
Original-Revision: 1d163074881af206037d46d5f1e3205919048307
GitOrigin-RevId: 618d0a55cf63f76a11a4aaf923fb28a67a09a6dc
Change-Id: I8480e306455545a46d842b1c900633b55a746a7e