[roll] Roll fuchsia [netdevice] Remove race possibilities
SessionStarted and SessionStopped paths were releasing and reacquiring
locks, which is breeding ground for races. Remove those with the help of
TA and letting `DeviceInterface` release the locks when it is done
changing the state of the data path.

We still release and reacquire the lock on NotifyDeadSession, which is
less concerning because the lock reacquisition is mostly to deal with
the zombie session, but the data path is already closed.

Multiply: network-device-lib-test: 30
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660423
Original-Revision: 26463c32da18aaa85b427d7a731b9bfef1fa50ab
GitOrigin-RevId: 6235e0723d5d7344e3798bd14b8a3dba2810e6f2
Change-Id: I67d339b2dcdc3eee09153bc814c641ce67d42d71