[roll] Roll fuchsia [usb][hub] fix unbind deadlock
Originally we were deadlocking because the completion event was never
being signaled when returning from RequestQueue in ControlOut on a
failure, resulting in a deadlock.

Here we also fix the problem of the thread trying to join even when the
thread is not joinable.

Original-Bug:94220

Test:Tested on Atlas

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657447
Original-Revision: 4af6907ba0a795fda5368634f74cfb2558371ae1
GitOrigin-RevId: d28101fdb248237424b813212011e0e984883490
Change-Id: I1099101f055686f8245818292ba519c244363112