[roll] Roll fuchsia [dhcpv6] Prepare Requesting for Reply to Renew
Currently the `Requesting` state holds `AddressToRequest` address
entries. In an upcoming change processing Reply to Renew, the client
will need to be able to transition to `Requesting` if it receives IAs
with status NoBinding, while also maintaining the existing
AddressEntry::Assigned entries for which the server had extended the
lifetime, as described in RFC 8415 section 18.2.10.1.

In preparation for implementing Reply to Renew, update the `Requesting`
state to hold `AddressEntry` addresses instead of `AddressToRequest`.
This is a mechanical change, `Requesting::addresses` still only contains
`AddressEntry::ToRequest` entries, it's just the type of the address map
that is updated at this stage.

Original-Bug: 76765

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654144
Original-Revision: 8e5cf1b61a182f2c92bc7f5af449de07bbe4e7a5
GitOrigin-RevId: 6d03228021f074aa264f4479cb0f4d896120408b
Change-Id: I2f2e40eec79d76f8781f8a610b289fd4bbd24302