[roll] Roll fuchsia [kernel][gfx-font] Remove dependency on libgfx-common
There are too many gfx-related libraries, and only one of them needs to
stand apart from the others at the moment: gfx-font-data (renamed in
this CL to gfx-font). Its sole job is to define some default font
configurations for both gigaboot and the kernel - and hence needs to
remain as pure C. Previously it depended on libgfx-common for a single,
simple struct definition, which in turn forced gfx-common to be written
as C (at least in part). So that we can better consolidate and improve
gfx code, this change reverse the dependency between these two
libraries, migrating the simple struct definition over.

Original-Bug: 96043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660814
Original-Revision: 8b8b1c49571e3a52f72cd1433c777299c0410493
GitOrigin-RevId: 0abb245925a66cbc6999732876729ca83d3435ba
Change-Id: I153cbb5fed4c6345b4e2e4f792fd1e61c28b9d41