[roll] Roll fuchsia [trace_processing] Fix check for missing 'fixed' events
The logic in the metric was intended to cleanly return an empty set of
metrics when the data was missing, but would actually throw an exception
in some cases.

In practice, this was never an issue previously as memory_monitor data
would start being collected as the trace started, so there was always a
data point present. With trace slicing, we may slice to a sub-1-second
interval that does not have any samples, so it is important to be able
to cleanly return in these cases.

TEST=fx test sl4f_client_tests
BUG: b/223316328

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660669
Original-Revision: ef71bead3ad8f68c8efee6b368c6780f30ed9ea2
GitOrigin-RevId: 0f0cdb12940ca929603889d65eebf7d48db1d3a4
Change-Id: I614db2b0b54004d1119bb73a48d5dd038d377841