[roll] Roll fuchsia [lockdep] No analysis on GuardMultiple::Release
The clang analysis is able to realize that the Guards being released in
GuardMultiple::Release are not held, however the annotation system is
not powerful enough to forward the lock requirements to the function
declaration. As such, we must disable the ananlysis instead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661102
Original-Revision: 7fe91401318d4826b2f3f78415f1bcfa8c719b5e
GitOrigin-RevId: f617e713b48108d5a4f757b9b831c823e5b35b0f
Change-Id: I0d0b93562e32970dcb8c352478f7012397417dbd