[roll] Roll fuchsia [setui] Raise camera device timeout
SetUI waits for a camera to be found for the input API on startup. If we
don't find a camera, we crash, which brings down the system. This is for
privacy reasons to ensure the user has the correct info for the privacy
state.

However, it seems that the 3 second timeout is not always enough as
there aren't any particular guarantees on the camera API side. Raising
the timeout to 30 seconds so that the system shouldn't crash in any
reasonable case.

Original-Bug: 96178
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661195
Original-Revision: 3391a0d56cc5765a4185d0090bdc1613b903d512
GitOrigin-RevId: aa9df06c90b4bac282b85ae62c5cd269d518926f
Change-Id: I714c972ea5e988e183d5fd4508728cc42c90402d