[roll] Roll fuchsia [realm_builder][rust] move contents of `new` module to top-level
The realm builder implementation was under the `new` module, as it
coexisted with the old implementation while the realm builder RFC was
implemented. Since the old implementation is deleted, move the new
RealmBuilder implementation to the top level of the library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660603
Original-Revision: 411fa9ee67f87627816dbec6eb9a0fbe51dd92f9
GitOrigin-RevId: a35330c8414a284d8df17290df3afdd1c7eddcc0
Change-Id: I58d298eec94949c9fcbe197c64fc42a8e3cb0f49