[roll] Roll infra/recipes [clang_ml_training] Use new size checker
The old size checker is going away in favor of the new one which runs
during the build.

The new size checker is already running as part of the build, so just
display the results as JSON instead of explicitly running a size
checker.

Original-Bug: 95221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/661005
Original-Revision: 23017928419cc23eefb0f4e11ad593f22670f27d
GitOrigin-RevId: 82fcc59cdc8887e2ed47cba9ff4a874785f7e6ee
Change-Id: Ic5361e123223a464c8bfaf612dab1a8ca0b817cb