[roll] Roll fuchsia [docs][input] link from input_drivers to driver inventory
Add a  link from the "Fuchsia input drivers" page (which
explains the input driver architecture) to the  "Fuchsia
hardware drivers" page (which enumerates in-tree drivers),
to make it easy to find the list of in-tree input drivers.

Note: I couldn't find a way to link to the filtered list, so
the link will show all drivers, not just input drivers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631741
Original-Revision: 7a5beb3ac5fd820bedb3043877772ada6a78d20f
GitOrigin-RevId: 4cbccb9428932315618b990384c2dc31e62387c2
Change-Id: Ic5704a75bada8c2b2ec82d6fabec3fb45497000b