[roll] Roll fuchsia [starnix] Add the concept of a "galaxy"
Prior to this change, each component that used Starnix as a runner would
specify how to configure the runtime environment. For example, a
component's .cml would contain a list of filesystem to mount, the kernel
command line, etc. The component would also contain the system image for
the environment in which it was to be run.

This change introduces the concept of a "galaxy." A galaxy is an
instance of the Starnix runner that is packaged with a system image and
a particular configuration (e.g., filesystems to mount, etc.). Each
such instance is exposed as a separate runner. It's important to note
that the starnix runner code and .cml is the same, only the system image
and configuration parameters change between galaxies.

For example, in a future change our gVisor tests will run using the
"stardroid_test" runner, and our chromiumos shell test will run using
the "starmiumos_test" runner (official name TBD).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644365
Original-Revision: 65398b2e301b4fca9ccbb5859a5f13a389fa46e2
GitOrigin-RevId: d043c695f1635c075073b722fa495a7459a6af18
Change-Id: Ida7cb332d2a1dd17ef486d4240954cde01ff0781