[roll] Roll fuchsia [fxfs] Add support for filtering an iterator
I started looking at the graveyard which had a TODO about using a filter
iterator and then noticed that the major compaction iterator was broken:
it didn't check the first item. It's not a big deal because I don't
think there's a case where we'd ever need to filter the first item, but
this should fix it.

I've chosen to leave the graveyard implementation alone since it's doing
more than just filtering.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661282
Original-Revision: 0c00eb1639355f546441124cc226de6eb55932e1
GitOrigin-RevId: b61fbd187d92186039f2420e19151083721fd771
Change-Id: I6644a55af48466a7a4882df8ad02ffc4565b28f9