[roll] Roll fuchsia [starnix][proc] Make writing proc directories easier
Abstracts away some of the details of offsets in `readdir` at the cost
of some throwaway computation (generating entries that will be discarded
due to the offset). This is a good trade-off that can be revisited in
the future once performance is deemed to be an issue.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661724
Original-Revision: 415a499d2e8338f5a0488c17e92fd0d9939da70e
GitOrigin-RevId: aa3acc55727d076f3ddbc07682403a8a4e3f00a0
Change-Id: I14c912ccca14c61aed9499bfd02647a3598f65f3