[roll] Roll fuchsia [run-test-suite] Don't check stdio for incomplete test cases
In the case that a test never sends a test case Stopped event,
or no status is contained in the Stopped event, run-test-suite
does not drain stdout and stderr. I argue this is correct - the
other end of the socket may be stuck open, which would cause
run-test-suite to hang if it tries to drain it. Instead, we'll
drop the socket, which can cause missing stdio. This is okay as
we report cases where Stopped is missing as a special error anyway,
so developers should still be able to diagnose the error.

This change updates the tests to not check for exact stdout output to
remove a test flake.

Multiply: run_test_suite_integration_tests
Original-Bug: 93130

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662185
Original-Revision: 2ec59893ddff97c6f1aa7f2341fd1cb41d4628f9
GitOrigin-RevId: 30c41a3b2f4c7b761ef116001e842599297112f4
Change-Id: I57dc9a25f092cc7d537950af6bfee2cc65654e4b