[roll] Roll fuchsia [ffx] Provide hint for finding host logs on internal errors
When ffx encounters an internal error, it cannot parse the error
and report guidance. When the error was caused by a plugin that
uses rust's logging facility, the best place for a developer to
look for more information is ffx host logs. However, the developer
may not know where ffx host logs are stored.

This change reports the host log directories to stderr (with a
fallback to stdout) when ffx encounters an internal error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659782
Original-Revision: 464b0346f84866b0f7ea513214110e7b184a6823
GitOrigin-RevId: 1294d0877fc677ce5e4f8cc80804a7268d9f1e64
Change-Id: I72fd0333282e2fba056a286a3e71f655c4c0d7f3