[roll] Roll fuchsia [FuSS] Add fuss_testing in-tree tests to allowlists required.
tqr/511573, as part of its build artifacts, creates a roughly minimal zbi
and blobfs in order to test that ffx scrutiny extract commandline
functionality works as FuSS (which is out-of-tree) expects it to.

This causes us to need to add it to particular allowlists.

b/222339195

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662231
Original-Revision: db7f89126aae643527c946309e562e8c54a5fa16
GitOrigin-RevId: b7b90c74eb9472bf2dba04688c854826a4b4dd68
Change-Id: I90f98b463435047a7183ed16ba273c67ab5f7754