[roll] Roll fuchsia [hypervisor][arm64] Remove superfluous lock
We previously had to initialise the VPID allocator, and therefore had to
hold a lock to do so. Since that code had been refactored to remove the
Init() function, we no longer need to hold the lock for the allocator.

Original-Bug: 95763
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662383
Original-Revision: 370cb86a6c046c6457089d3c4b524fe20db7311d
GitOrigin-RevId: 1bd518d5a2e9545804586edbe9ab087c79751e85
Change-Id: I6928ebd6bf88af463aedee77e8b2ff71745a64ef