[roll] Roll fuchsia [audio][test] Move ConfigurePipelineForThermal to parent
For the most part, support for testing audio with the thermal-
mitigation API is defined in HermeticAudioTest, a parent to
HermeticFidelityTest. One method ConfigurePipelineForThermal
is defined in HermeticFidelityTest but is not specific to
the requirements of that test. This CL moves the method from
HermeticFidelityTest to HermeticAudioTest.

Although not at all required for it, this is being done as
part of the work to integrate the new thermal configuration
API.

fx test -do audio_core_google_comms_thermal_tests_astro

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662486
Original-Revision: 625d1ea22cd955c7503f10fc907fdad0f093ce97
GitOrigin-RevId: 4d30d92fa62ad873c4b80d138b4ac0fabd6b6084
Change-Id: If65f5e98d99d6f75a6dfc182a321ccb8343eb5a2