[roll] Roll fuchsia [fdf] Fix crash in CompositeDevice::BindFragment
With the new workflow for the driver index, it is
possible to attempt to bind a fragment that's
already bound in CompositeDevice. This CL fixes the
issue by changing the behavior to return ZX_OK and
log an error.

Original-Bug: 96220
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661463
Original-Revision: a290e36c3638e30246ba2d4233bc728b3cb70c43
GitOrigin-RevId: bc5df6af200dcb4b0ec352dc39081f7e36354755
Change-Id: Iabd7b6ea19e0e92a30a7914937c6633b28d0befa