[roll] Roll fuchsia [fdf] Driver manager supports device groups
Implement support for device groups in driver
manager. The Coordinator adds the device group by
first adding it to the driver index through the
driver loader, before passing the device group to
the BindDriverManager for binding the nodes.

The BindDriverManager creates a CompositeDevice
for the device group and maps it according to the
topological path.

Original-Bug: 91510
Tests: Add unit test to driver-manager-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659263
Original-Revision: 62174108e02c85feb7a18df5cc03dcf8ec7d8625
GitOrigin-RevId: fd48d5f0ada51bbdb76435d63118d3bfefa6981a
Change-Id: If41ad64c92e7e8e07cc3fcd02f75e6595f6a2048