[roll] Roll fuchsia [assembly] Remove the image assembly config input
Now that we have migrated over to product assembly, we no longer need to
upload the image assembly config and it's dependencies, because the
config is generated by the product assembler, and the dependencies are
included in the legacy AIB.

We also include the product assembly config so that it can be update
in-tree and automatically added out-of-tree.

Unfortunately, we cannot simply this to a generated_file(), because we
need to add the custom signing script and the vbmeta keys. In order to
add these files, we have to strip the ../../ prefix, which I do not
believe we can easily do in GN.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662804
Original-Revision: 124dc72e50c44e081683a42db3a7df612e889931
GitOrigin-RevId: bfdbb81e9776eec97184de35c212dea61ea0fd6f
Change-Id: Iaca10488350c843c928812d89a2639f0ead8a53f