[roll] Roll fuchsia [lsc] Migrate all clients of fuchsia.sys2.ComponentResolver
This LSC migrates all usages of the `fuchsia.sys2.ComponentResolver`
protocol to `fuchsia.component.resolution.Resolver`. The latter is in
the SDK and the former will be removed after this CL. This was done
in an LSC, instead of a soft transition for the following reasons:

* All usages are in-tree, so we're confident that nothing will break.
* The change is mechanical. That is, it only changes one namespace to
another, and as such it's fair to assume that nothing will break.
* A soft transition would have greatly slowed down the rollout
of `fuchsia.component.resolution.Resolver`.
* It would have been tricky to have `component_manager` support
consuming both protocols at the same time.

Original-Bug: 94581

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656983
Original-Revision: 28e39762b3aaec2c3d3bf4d1b29fbb87ca1e1848
GitOrigin-RevId: a31efe32d3846c26dd249d9fae590988061f5576
Change-Id: I0bfe4e3c1c6dc6c1d0b946236a77c91adc27dc5c