[roll] Roll fuchsia [fshost] Initialize crypt keys on boot
When the system is configured to use Fxfs, initialize the crypt service
with some dummy keys on boot.  This is necessary for using actual crypto
algorithms in Crypt.

The keys themselves are not secure and not intended to be.  A later CL
will switch to reading from a hardware source.

Original-Bug: 94587
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656493
Original-Revision: 9a22599ce6b364da27f00bb7299fc311b0017097
GitOrigin-RevId: df3a6c8c28f7c71e47598190a1ff1448f1e0d76a
Change-Id: I62cfd02e996e60986c152d4e62a9c5433b32d54b